% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_missing_entries.R
\name{add_missing_entries}
\alias{add_missing_entries}
\title{Add missing entries}
\usage{
add_missing_entries(df, info, id, field)
}
\arguments{
\item{df}{data.frame with the fields given in the 'id' and 'field' arguments}

\item{info}{Information table including the 'id' and 'field' fields}

\item{id}{Field that is used to match 'df' and 'info' data.frames given as input}

\item{field}{Field that will be filled in 'df' based on the information in 'info' data.frame}
}
\value{
Character vector of unique author IDs
}
\description{
Fill in missing entries in the data based on supplied information table
}
\examples{
\dontrun{a <- add_missing_entries(...)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
