% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_years.R
\name{approximate_years}
\alias{approximate_years}
\title{Approximate Years}
\usage{
approximate_years(x)
}
\arguments{
\item{x}{year field (a vector)}

\item{start_synonyms}{Synonyme table for start year}

\item{end_synonyms}{Synonyme table for end year}

\item{verbose}{verbose}

\item{min.year}{Minimum year accepted}

\item{max.year}{Maximum year accepted}

\item{check}{If true, remove entries (replace by NA) where start > end}
}
\value{
data.frame with the fields 'start' and 'end'
}
\description{
Approximate year for entries that are approximations.
}
\examples{
\dontrun{df <- polish_years(c("1746", "1745-1750"))}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
