% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_unique.R
\name{author_unique}
\alias{author_unique}
\title{author_unique}
\usage{
author_unique(df)
}
\arguments{
\item{df}{data.frame with fields "author_name", "author_birth", "author_death"}
}
\value{
Character vector of unique author IDs
}
\description{
Form unique author identifiers by combining name and life years
}
\examples{
\dontrun{a <- author_unique(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

