% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{build_pattern}
\alias{build_pattern}
\title{Build the pattern for recognizing gathering sets}
\usage{
build_pattern()
}
\value{
pattern (a regular expression)
}
\description{
Build the pattern for recognizing gathering sets
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{internal}
