% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{build_signature_table}
\alias{build_signature_table}
\title{Build the signature table}
\usage{
build_signature_table()
}
\value{
signature table
}
\description{
This builds a list of possible gathering signatures in their
sequential ordering. It is necessary to represent this table explicitly when
unfolding sequences of gatherings, because the meaning of sequences is far
from obvious.
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{internal}
