% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{capitalize}
\usage{
capitalize(x)
}
\arguments{
\item{x}{a character vector to capitalize}
}
\value{
Capitalized character vector
}
\description{
Capitalize the first letter
}
\examples{
# capitalize("print")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

