% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{century}
\alias{century}
\title{Get Century}
\usage{
century(x)
}
\arguments{
\item{x}{a numeric variable}
}
\value{
Century. A numeric.
}
\description{
Round year into century.
}
\examples{
century(1962)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
