% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cheat_publishers.R
\name{cheat_publishers}
\alias{cheat_publishers}
\title{Cheat Publishers}
\usage{
cheat_publishers()
}
\value{
Data frame with possible publisher form and the preferred output
}
\description{
Get publisher list from an xml file.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}

