% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_comhis.R
\name{check.packages}
\alias{check.packages}
\title{Theme function for COMHIS group}
\usage{
check.packages(pkg)
}
\arguments{
\item{type}{Type of the data to be plotted ("continuous"; "discrete")}
}
\description{
default theme for consistent plot style within the 
COMHIS group.
}
\author{
Abhishekh Gupta \email{abishak.gupta@gmail.com}
Usage: source("theme_comhis.R") and append '+ theme_comhis(type)' 
to the desired plot
}
