% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/synonymes.R
\name{check_synonymes}
\alias{check_synonymes}
\title{check_synonymes}
\usage{
check_synonymes(synonymes)
}
\arguments{
\item{synonymes}{synonymes data.frame}
}
\value{
Polished synonyme table
}
\description{
Check synonyme table.
}
\details{
Remove duplicated information. Ensure identical matches
are included in synonyme list.
}
\examples{
\dontrun{s <- check_synonymes(synonymes)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

