% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/volume.R
\name{check_volumes}
\alias{check_volumes}
\title{check_volumes}
\usage{
check_volumes(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Volume information
}
\description{
Pick volume information from page number field
}
\examples{
check_volumes("v.4, 293")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

