% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_publisher_destructively.R
\name{clean_publisher_destructively}
\alias{clean_publisher_destructively}
\title{clean publisher destructively}
\usage{
clean_publisher_destructively(x, languages = c("english"))
}
\arguments{
\item{x}{A vector of publisher names}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Vector of publisher names
}
\description{
Highly data specific cleansing of publisher field into a temporary form
}
\examples{
# clean_publisher(x, languages=c("finnish", "swedish", "latin"))
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

