% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_table.R
\name{cleanup_synonyme_table}
\alias{cleanup_synonyme_table}
\title{cleanup_synonyme_table}
\usage{
cleanup_synonyme_table(f, sep = ";", write = FALSE)
}
\arguments{
\item{f}{File name}

\item{sep}{Separator}

\item{write}{Write back to file after cleaning up ? Logical.}
}
\value{
Polished synonyme table.
}
\description{
Clean up the synonyme table.
}
\details{
Adds, lowercase version of each name; removes duplicates;
         writes back to file (optional)
}
\examples{
# tmp <- cleanup_synonyme_table(f, sep = ";", write = FALSE)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

