% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_title_paper.R
\name{compare_title_paper}
\alias{compare_title_paper}
\title{Compare Title Count and Paper Consumption}
\usage{
compare_title_paper(x, field, selected = NULL, plot.mode = "text")
}
\arguments{
\item{x}{data frame}

\item{field}{Field to analyze}

\item{selected}{Limit the analysis on selected entries}

\item{plot.mode}{"text" or "point"}
}
\value{
List:
\itemize{
  \item{plot}{ggplot object}
  \item{table}{summary table}
}
}
\description{
Compare title count and paper consumption for selected field.
}
\examples{
\dontrun{compare_title_paper(df, "author")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
