% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_field.R
\name{compress_field}
\alias{compress_field}
\title{Compress Field into Fewer Categories}
\usage{
compress_field(x, keep, rest = "Other")
}
\arguments{
\item{x}{Vector or factor.}

\item{keep}{Number of top entries to include; or a character vector}

\item{rest}{Name to be used for the remaining entries}
}
\value{
Compressed version of the input.
}
\description{
Merge the less common terms into a single group
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
