% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round.R
\name{decade}
\alias{decade}
\title{Get Decade}
\usage{
decade(x)
}
\arguments{
\item{x}{a numeric variable}
}
\value{
Decade. A numeric.
}
\description{
Round year into decade.
}
\examples{
decade(1962)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
