% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decapitate_keywords.R
\name{decapitate_keywords}
\alias{decapitate_keywords}
\title{Decapitate keywords}
\usage{
decapitate_keywords(x, languages = c("english"))
}
\arguments{
\item{x}{Vector of personal names}

\item{languages}{A vector of languages which are used in detecting keywords}
}
\value{
Vector of words
}
\description{
Removes keywords
}
\examples{
# extract_personal_names(x, languages=c("finnish", "swedish", "latin"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
