% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{dimension_table}
\alias{dimension_table}
\title{dimension_table}
\usage{
dimension_table(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Document dimension table
}
\description{
Document dimension mapping table
}
\examples{
dimension_table()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

