% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_dimensions.R
\name{enrich_dimensions}
\alias{enrich_dimensions}
\title{Enrich Dimension Field}
\usage{
enrich_dimensions(df)
}
\arguments{
\item{df}{Preprocessed data.frame}
}
\value{
Augmented data.frame
}
\description{
Augment dimensions.
}
\examples{
\dontrun{df2 <- enrich_dimensions(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
