% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_geo.R
\name{enrich_geo}
\alias{enrich_geo}
\title{Enrich Geodata}
\usage{
enrich_geo(df)
}
\arguments{
\item{df}{Preprocessed data.frame}
}
\value{
Augmented data.frame
}
\description{
Enrich geodata.
}
\examples{
\dontrun{df2 <- enrich_geo(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

