% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_pagecount.R
\name{enrich_pagecount}
\alias{enrich_pagecount}
\title{Enrich Page Count Field}
\usage{
enrich_pagecount(df)
}
\arguments{
\item{df}{Preprocessed data.frame}
}
\value{
Augmented data.frame
}
\description{
Augment missing pagecounts based on mean estimates from available data.
}
\examples{
\dontrun{df2 <- enrich_pagecount(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

