% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_years.R
\name{enrich_years}
\alias{enrich_years}
\title{Enrich Years}
\usage{
enrich_years(df.preprocessed, df.orig)
}
\arguments{
\item{df.preprocessed}{Preprocessed data.frame}

\item{df.orig}{Original data.frame}
}
\value{
Augmented data.frame
}
\description{
Enrich year info.
}
\examples{
\dontrun{df2 <- enrich_years(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
