% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pages_univol.R
\name{estimate_pages_singlevol}
\alias{estimate_pages_singlevol}
\title{Page Estimation (Single Volume)}
\usage{
estimate_pages_singlevol(df, mean.pagecounts.singlevol)
}
\arguments{
\item{df}{data.frame of documents x variables}

\item{mean.pagecounts.singlevol}{Page count estimates to be used}
}
\value{
Page count estimates
}
\description{
Estimate pages for single volume documents.
}
\examples{
\dontrun{p <- estimate_pages_singlevol(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

