% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pages_univol.R
\name{estimate_pages_univol}
\alias{estimate_pages_univol}
\title{Page estimation for univols}
\usage{
estimate_pages_univol(df, mean.pagecounts.univol)
}
\arguments{
\item{df}{data.frame of documents x variables}

\item{mean.pagecounts.univol}{Page count estimates to be used}
}
\value{
Page count estimates
}
\description{
Estimate pages for univols.
}
\examples{
\dontrun{p <- estimate_pages_univol(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

