% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{eval_collation_formula}
\alias{eval_collation_formula}
\title{Evaluate collation formula}
\usage{
eval_collation_formula(x)
}
\arguments{
\item{x}{input char}
}
\value{
vector containing numbers of leaves
}
\description{
Evaluates a collation formula to the number of leaves.
}
\examples{
eval_collation_formula("n1 A`SUP`8`LO` B`SUP`6`LO`")
}
\references{
See citation("bibliographica") and explanations in
            \url{https://manual.stcv.be/p/Collation_Formula}.
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{utilities}
