% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_personal_names.R
\name{extract_personal_names}
\alias{extract_personal_names}
\title{Extract personal names}
\usage{
extract_personal_names(x, languages = c("english"))
}
\arguments{
\item{x}{Vector of publisher names}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Data frame: orig, initials, family, full_name, init_name, guessed, relation
}
\description{
Extracts personal names from
}
\examples{
# extract_personal_names(x, languages=c("finnish", "swedish", "latin"))
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

