% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_availability.R
\name{field_availability}
\alias{field_availability}
\title{Field availability}
\usage{
field_availability(df)
}
\arguments{
\item{df}{data.frame to explore}
}
\value{
List with following elements
\itemize{
  \item{plot}{ggplot2 object}
  \item{table}{Table summarizing the field name (field_name), overall fraction of available (non-NA) and missing entries (NAs) as percentages [0, 100], number of non-NA entries (n), and the number of unique entries (unique_entries)}
}
}
\description{
Estimate fraction of missing entries (NAs) and visualize data availability for the different fields.
}
\examples{
# field_availability(df)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
