% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_table.R
\name{field_table}
\alias{field_table}
\title{Field Table}
\usage{
field_table(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Field codes table
}
\description{
Field codes mapping table.
}
\examples{
field_table()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
