% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fieldname_table.R
\name{fieldname_table}
\alias{fieldname_table}
\title{Field Name Table}
\usage{
fieldname_table(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Field names table
}
\description{
Field names mapping table.
}
\examples{
fieldname_table()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
