% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fill_dimensions.R
\name{fill_dimensions}
\alias{fill_dimensions}
\title{fill_dimensions}
\usage{
fill_dimensions(x, dimension.table)
}
\arguments{
\item{x}{dimension string}

\item{dimension.table}{Given mappings between document dimensions}
}
\value{
Augmented dimension vector
}
\description{
Estimate missing entries in dimension vector where possible
}
\examples{
# dimension.table <- dimension_table();
          #  fill_dimensions(c(original = NA, gatherings = NA,
	     #			  width = 30, height = 48),
	     #			  dimension.table)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\seealso{
augment_dimension_table
}
\keyword{utilities}

