% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dissertations.R
\name{filter_dissertations}
\alias{filter_dissertations}
\title{Filter dissertations}
\usage{
filter_dissertations(df.combined, catalog_name = "ANY", town = "ANY")
}
\arguments{
\item{df.combined}{Data frame of processed data from both catalogs}

\item{catalog_name}{Name of catalog inspect. Default "ANY"}

\item{town}{Publication place. Default "ANY"}
}
\value{
Filtered data frame
}
\description{
Gets dissertations from a single catalog or all of them
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
