% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicates.R
\name{filter_duplicates}
\alias{filter_duplicates}
\title{Filter duplicates}
\usage{
filter_duplicates(df.preprocessed, field_names = c("short_title",
  "publication_place", "publication_year"))
}
\arguments{
\item{field_names}{List of field names of exactly matching values}

\item{df.prepocessed}{Data frame of combined catalogs}
}
\value{
Data frame
}
\description{
Filters duplicates from df.preprocessed
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

