% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_towns.R
\name{filter_towns}
\alias{filter_towns}
\title{Filter towns}
\usage{
filter_towns(df.combined, catalog_name = "ANY", town = "ANY")
}
\arguments{
\item{df.combined}{Data frame of processed data from both catalogs}

\item{catalog_name}{Name of catalog to inspect. Default "ANY"}

\item{town}{Publication place. Default "ANY"}
}
\value{
Filtered data frame
}
\description{
Gets publications printed in a single town from a single catalog or all of them
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
