% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_custom}
\alias{firstnames_custom}
\title{Custom First Names}
\usage{
firstnames_custom(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
Custom first name table, including gender info.
}
\examples{
\dontrun{x <- firstnames_custom()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

