% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_finnish}
\alias{firstnames_finnish}
\title{Finnish First Names}
\usage{
firstnames_finnish(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
Finnish first name table, including gender info.
}
\details{
Finnish first names (including gender info) from Vaestorekisterikeskus VRK (see fennica::get_gender_fi) for details.
}
\examples{
\dontrun{x <- firstnames_finnish()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\seealso{
fennica::get_gender_fi
}
\keyword{utilities}

