% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_french}
\alias{firstnames_french}
\title{French first names}
\usage{
firstnames_french(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
French first names table, including gender info.
}
\details{
The French first name lists are collected from \url{http://www.excel-downloads.com/forum/86934-liste-des-prenoms.htmlhttp://http://www.excel-downloads.com/forum/86934-liste-des-prenoms.html}.
}
\examples{
\dontrun{x <- firstnames_french()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

