% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_german}
\alias{firstnames_german}
\title{German first names}
\usage{
firstnames_german(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
German first name table, including gender info.
}
\details{
The information was collected from \url{http://www.albertmartin.de/vornamen/}.
}
\examples{
\dontrun{x <- firstnames_german()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

