% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_multilingual}
\alias{firstnames_multilingual}
\title{Multilingual first names}
\usage{
firstnames_multilingual(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
Multilingual first names table, including gender info.
}
\details{
Source: multilingual \url{http://www.lexique.org/public/prenoms.php} (Prenoms.txt).
}
\examples{
\dontrun{x <- firstnames_multilingual()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

