% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames_pseudo}
\alias{firstnames_pseudo}
\title{Pseudonyme Information}
\usage{
firstnames_pseudo(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name (pseudonyme) and gender info.
}
\description{
Pseudonyme first names, including gender info.
}
\examples{
\dontrun{x <- firstnames_pseudo()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

