% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_period.R
\name{format_period}
\alias{format_period}
\title{Format Time Periods}
\usage{
format_period(x)
}
\arguments{
\item{x}{A vector (numeric/character/factor)}
}
\value{
Converted entries.
}
\description{
Conversions between time period formats.
}
\details{
Converts format [1700, 1800) into 1700-1799. Used in polishing tables and figures.
}
\examples{
format_period("[1700, 1800)")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
