% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gatherings_size.R
\name{gatherings_size}
\alias{gatherings_size}
\title{Gatherings Size}
\usage{
gatherings_size(x, sheet.dimensions = NULL)
}
\arguments{
\item{x}{Gatherings name}

\item{sheet.dimensions}{Mapping between document dimensions.}
}
\value{
Vector (single input) or matrix (multiple inputs)
}
\description{
Return size of a given document format (gatherings) in cm.
}
\examples{
d <- gatherings_size("Octavo")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
