% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gatherings_table.R
\name{gatherings_table}
\alias{gatherings_table}
\title{gatherings_table}
\usage{
gatherings_table(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Document gatherings table
}
\description{
Document gatherings names mapping table
}
\examples{
gatherings_table()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

