% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gender_custom.R
\name{gender_custom}
\alias{gender_custom}
\title{Custom Name-Gender Mappings}
\usage{
gender_custom(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Table with first name and gender info.
}
\description{
Custom first name table, including gender info.
}
\examples{
\dontrun{x <- gender_custom()}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
