% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gendermap}
\alias{gendermap}
\title{Name - Gender Mapping Table}
\format{data.frame}
\usage{
data(gendermap)
}
\description{
Mappings between first names and gender, collected from various sources.
}
\details{
The steps to reproduce this table, and references to the original data sources, are listed in the inst/extdata/data.gender.R function of the bibliographica package.
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{data}

