% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_tables.R
\name{generate_summary_tables}
\alias{generate_summary_tables}
\title{Generate summary tables}
\usage{
generate_summary_tables(df.preprocessed, df.orig,
  output.folder = "output.tables",
  discarded.author.table = discarded.author.table,
  discarded.author.firstnames = discarded.author.firstnames,
  discarded.author.lastnames = discarded.author.lastnames)
}
\arguments{
\item{df.preprocessed}{Preprocessed data.frame to be summarized}

\item{df.orig}{Original data.frame for comparisons}

\item{output.folder}{Output folder path}

\item{discarded.author.table}{discarded.author.table}

\item{discarded.author.firstnames}{discarded.author.firstnames}

\item{discarded.author.lastnames}{discarded.author.lastnames}
}
\description{
Generate summary tables from the preprocessed data frame.
}
\examples{
# generate_summary_tables(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

