% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summary_tables_geo.R
\name{generate_summary_tables_geo}
\alias{generate_summary_tables_geo}
\title{Summary Tables for Geodata}
\usage{
generate_summary_tables_geo(
  df.preprocessed,
  df.orig,
  output.folder = "output.tables"
)
}
\arguments{
\item{df.preprocessed}{Preprocessed data.frame to be summarized}

\item{df.orig}{Original data.frame for comparisons}

\item{output.folder}{Output folder path}
}
\description{
Generate summary tables from the preprocessed data frame.
}
\examples{
# generate_summary_tables_geo(df)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
