% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{get_country}
\usage{
get_country(x, map = NULL)
}
\arguments{
\item{x}{A vector of region names (cities or municipalities etc.)}

\item{map}{data.frame with region to country mappings
               (fields 'region' and 'country')}
}
\value{
data.frame
}
\description{
Map geographic places to country names
}
\examples{
x2 <- get_country("Berlin")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

