% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dissertation_mismatches.R
\name{get_dissertation_mismatches}
\alias{get_dissertation_mismatches}
\title{Get dissertation mismatches}
\usage{
get_dissertation_mismatches(combined_catalogs)
}
\arguments{
\item{combined_catalogs}{Data frame of combined catalogs}
}
\value{
Data frame of core fields
}
\description{
Lists titles which are tagged as dissertations in another catalog, but not in the other one
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
