% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dissertations.R
\name{get_dissertations}
\alias{get_dissertations}
\title{Get dissertations}
\usage{
get_dissertations(processed_frame, catalog_name = "ANY", before = NULL)
}
\arguments{
\item{processed_frame}{Data frame of processed data}

\item{catalog_name}{Name of catalog inspect. Default "ANY"}

\item{before}{Upper limit of publication year (self included)}
}
\value{
Data frame of core fields
}
\description{
Gets dissertations from a single catalog or all of them
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
