% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gender.R
\name{get_gender}
\alias{get_gender}
\title{Pick gender based on first names}
\usage{
get_gender(x)
}
\arguments{
\item{x}{Vector of first names}
}
\value{
Author gender information
}
\description{
Pick gender based on first names
}
\examples{
get_gender("julius")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

