% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geocoordinates.R
\name{get_geocoordinates}
\alias{get_geocoordinates}
\title{Get Geocoordinates}
\usage{
get_geocoordinates(x, geonames, places.geonames)
}
\arguments{
\item{x}{A vector of publication place names}

\item{geonames}{geonames}

\item{places.geonames}{places.geonames}
}
\value{
data.frame with latitude and longitude
}
\description{
Map geographic places to geocoordinates.
}
\details{
Experimental.
}
\examples{
\dontrun{x2 <- get_geocoordinates("Berlin")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
