% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_pagecounts.R
\name{get_mean_pagecounts}
\alias{get_mean_pagecounts}
\title{Mean Page Counts for All Document Types}
\usage{
get_mean_pagecounts(df, exclude.plates = TRUE)
}
\arguments{
\item{df}{data frame, including the required fields.}

\item{exclude.plates}{Exclude plate info from page counts.}
}
\value{
Average page count information.
}
\description{
Calculate mean page counts for different document types.
}
\examples{
\dontrun{m <- get_mean_pagecounts(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
