% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publishers_Finto.R
\name{get_publishers_Finto}
\alias{get_publishers_Finto}
\title{Get publishers Finto}
\usage{
get_publishers_Finto(Finto_corrected, Finto_comp, all_names, known_inds,
  Finto_town, unknown_town, publication_year, Finto_years)
}
\arguments{
\item{Finto_corrected}{Vector of preferred names from Finto}

\item{Finto_comp}{Vector of alternative names from Finto}

\item{all_names}{Data frame containing all the name related fields}

\item{known_inds}{Vector of already processed indices, to ignore}

\item{Finto_town}{Vector of towns in Finto}

\item{unknown_town}{Vector of towns from raw data}

\item{publication_year}{Data frame of published_in, published_from and published_till}

\item{Finto_years}{Data frame of published_in, published_from and publishe_till}
}
\value{
Data frame with alt, pref and match_methods
}
\description{
Use Finto database to unify publisher names
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

