% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_missing_entries.R
\name{guess_missing_entries}
\alias{guess_missing_entries}
\title{guess_missing_entries}
\usage{
guess_missing_entries(id, values)
}
\arguments{
\item{id}{identifier vector}

\item{values}{corresponding values with potentially missing information (NAs)}
}
\value{
A vector with augmented values
}
\description{
Fill in the missing values. This function assumes that each id has a unique value which is missing for some occurrences of the id but available in others and copied from there to fill in the missing occurrences.
}
\details{
For instance, we may have authors and author life years (birth and death).
The life years may be available for a given author in some entries and missing in others.
When the information is unique, this function fills the missing entries where possible.
The function checks that the available life years for the given id (author in this
example) are unique (ie. to avoid ambiguous mappings, for instance distinct authors with
identical name but different birth year). For unique id-value relation, use the unique
value to fill in the missing entries for the same id (
}
\examples{
\dontrun{guess_missing_entries(id = c("Tom", "Tom", "Pete", "Pete", "Pete"), values = c(1, NA, 2, 3, NA))}
}
\keyword{utils}

