% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_ie.R
\name{handle_ie}
\alias{handle_ie}
\title{handle_ie}
\usage{
handle_ie(x, harmonize = TRUE)
}
\arguments{
\item{x}{A vector}

\item{harmonize}{Logical. Harmonize ie statements efore interpretation?}
}
\value{
A vector polished
}
\description{
Handle ie statement
}
\examples{
\dontrun{x2 <- handle_ie("i.e.")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

