% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_abbreviated_names.R
\name{harmonize_abbreviated_names}
\alias{harmonize_abbreviated_names}
\title{Harmonize abbreviated names}
\usage{
harmonize_abbreviated_names(x, languages = "english")
}
\arguments{
\item{x}{A vector of publisher names}

\item{languages}{Language which are used in writing out the abbreviated names}
}
\value{
Vector of publishers names
}
\description{
Writes out abbreviated names.
}
\examples{
# harmonize_abbreviated_names(c("Phil.", "Joh."), languages=c("finnish"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
