% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{harmonize_dimension}
\alias{harmonize_dimension}
\title{harmonize_dimension}
\usage{
harmonize_dimension(x, sheetsizes)
}
\arguments{
\item{x}{A character vector that may contain dimension information}
}
\value{
The character vector with dimension information harmonized
}
\description{
Harmonize dimension information
}
\examples{
# harmonize_dimension("4to")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{internal}

