% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonize_field_names.R
\name{harmonize_field_names}
\alias{harmonize_field_names}
\title{Convert field codes to names}
\usage{
harmonize_field_names(x)
}
\arguments{
\item{x}{a vector of field codes}
}
\value{
Vector of field names
}
\description{
Convert field codes to names
}
\examples{
x2 <- harmonize_field_names("041a")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

