% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_for_comparison.R
\name{harmonize_for_comparison}
\alias{harmonize_for_comparison}
\title{Harmonize for Comparison}
\usage{
harmonize_for_comparison(x, languages = "english")
}
\arguments{
\item{x}{A vector of publisher names}

\item{languages}{languages used to detect orthographical variants}
}
\value{
Vector of publishers names
}
\description{
Harmonize orthography
}
\examples{
# harmonize_for_comparison(x, languages = "finnish")
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
