% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonize_names.R
\name{harmonize_names}
\alias{harmonize_names}
\title{harmonize_names}
\usage{
harmonize_names(x, synonymes)
}
\arguments{
\item{x}{A character vector}

\item{synonymes}{synonyme table.
       The selected names are in column 1, the synonymes in column 2.}
}
\value{
Harmonized vector where synonymes are renamed by the selected names
}
\description{
Harmonize names
}
\examples{
\dontrun{x2 <- harmonize_names(x, file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

