% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_print_statements.R
\name{harmonize_print_statements}
\alias{harmonize_print_statements}
\title{harmonize_print_statements}
\usage{
harmonize_print_statements(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Harmonized vector
}
\description{
Harmonize print statements
}
\examples{
# x2 <- harmonize_print_statements("prentyd in London")$name
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

