% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher.R
\name{harmonize_publisher}
\alias{harmonize_publisher}
\title{Harmonize Publisher}
\usage{
harmonize_publisher(df, languages = c("english"))
}
\arguments{
\item{df}{bibliographic data.frame with the following fields: publication_year, publisher}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Polished vector.
}
\description{
Harmonizes publishers with slight misspellings.
}
\examples{
\dontrun{harmonize_publisher(x, publication_year,
                     languages = c("finnish", "swedish", "latin"))}
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

