% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher_generic.R
\name{harmonize_publisher_generic}
\alias{harmonize_publisher_generic}
\title{Harmonize Publisher Generic}
\usage{
harmonize_publisher_generic(df.orig, languages = "english")
}
\arguments{
\item{df.orig}{Data frame with raw data, assuming the place, year fields are already polished.}

\item{languages}{languages to consider in polishing}
}
\value{
A vector with polished entries.
}
\description{
Main handler for publisher fields.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

