% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher_main.R
\name{harmonize_publisher_main}
\alias{harmonize_publisher_main}
\title{Harmonize Publisher Main}
\usage{
harmonize_publisher_main(datasource, df.orig, testing_max = NULL)
}
\arguments{
\item{datasource}{String denoting catalog: "fennica", "kungliga"...}

\item{df.orig}{Data frame with raw data}

\item{testing_max}{Integer 1:testing_max rows will be handled. For testing purposes only}
}
\value{
Data frame with orig, mod
}
\description{
Main handler for publisher fields.
}
\examples{
# clean_publisher(x, languages=c("finnish", "swedish", "latin"))
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

