% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publishers_per_language.R
\name{harmonize_publishers_per_language}
\alias{harmonize_publishers_per_language}
\title{Harmonize Publishers per Language}
\usage{
harmonize_publishers_per_language(q, languages)
}
\arguments{
\item{q}{Character vector}

\item{languages}{languages}
}
\value{
Harmonized terms
}
\description{
Harmonize publishers per language.
}
\examples{
\dontrun{x <- harmonize_publishers_per_language(q, languages)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
