% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.multivol.R
\name{is.multivol}
\alias{is.multivol}
\title{Identify multi volume docs}
\usage{
is.multivol(df)
}
\arguments{
\item{df}{data.frame of documents x variables}
}
\value{
Logical vector indicating the multi-volume docs
}
\description{
Identify documents that can be considered multi-volume; based on other document info.
}
\details{
This function is used only to estimate pagecounts for documents with missing page count information.
Therefore no page count is considered in assessing the issue status.
}
\examples{
\dontrun{issue <- is.multivol(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

