% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.singlevol.R
\name{is.singlevol}
\alias{is.singlevol}
\title{Identify single volume docs}
\usage{
is.singlevol(df)
}
\arguments{
\item{df}{data.frame of documents x variables}
}
\value{
Logical vector indicating the single-volume docs
}
\description{
Identify documents that can be considered single-volume; based on other document info.
}
\details{
This function is used only to estimate pagecounts for documents with missing page count information.
Therefore no page count is considered in assessing the issue status.
}
\examples{
\dontrun{issue <- is.singlevol(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
