% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lastnames.R
\name{lastnames}
\alias{lastnames}
\title{lastnames}
\usage{
lastnames(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Vector of last names
}
\description{
Vector of last names
}
\examples{
\dontrun{x <- lastnames()}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
