% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_comparisons.R
\name{list_comparisons}
\alias{list_comparisons}
\title{List comparisons}
\usage{
list_comparisons(combined, field_names = c("short", "town", "from"))
}
\arguments{
\item{combined}{Data frame of both catalogs}

\item{field_names}{List of field names of exactly matching values}
}
\value{
Data frame of core fields listed by their title
}
\description{
Lists similar titles from Fennica and Kungliga catalogs
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

