% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_initial_datafile.R
\name{load_initial_datafile}
\alias{load_initial_datafile}
\title{Load Intial Datafile}
\usage{
load_initial_datafile(
  datafile,
  reload.data = FALSE,
  output.file = "parsed.Rds",
  sep = "|"
)
}
\arguments{
\item{datafile}{Datafile in csv.gz format}

\item{reload.data}{Boolean, edfault FALSE. Use datafile or previously saved Rds?}

\item{output.file}{Output file for storing the resulting object.}

\item{sep}{separator}
}
\value{
data.frame with raw data fields
}
\description{
Load datafile in csv.gz format, or use previously saved Rds.
}
\examples{
\dontrun{df.orig <- load_initial_datafile(datafile, reload.data = TRUE)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti, Ville Vaara
}
\keyword{utilities}
