% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_initial_datafile_dev.R
\name{load_initial_datafile_dev}
\alias{load_initial_datafile_dev}
\title{Load Intial Datafile}
\usage{
load_initial_datafile_dev(datafile, ignore.fields, reload.data = FALSE)
}
\arguments{
\item{datafile}{Datafile in csv.gz format}

\item{ignore.fields}{Fields to ignore}

\item{reload.data}{Boolean, edfault FALSE. Use datafile or previously saved Rds?}
}
\value{
data.frame with raw data fields
}
\description{
Load datafile in csv.gx format, or use previously saved Rds.
}
\examples{
\dontrun{df.orig <- load_initial_datafile(datafile, ignore.fields, reload.data = TRUE)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti, Ville Vaara
}
\keyword{utilities}
