% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_fieldnames.R
\name{map_fieldnames}
\alias{map_fieldnames}
\title{Polish Field Names}
\usage{
map_fieldnames(x, from = "Data", to = "Name")
}
\arguments{
\item{x}{Character vector}

\item{from}{Field to map from.}

\item{to}{Field to map to.}
}
\value{
Mapped terms
}
\description{
Convert the field names in the data in print-friendly form.
}
\examples{
map_fieldnames("gatherings") 
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
