% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_gatherings.R
\name{map_gatherings}
\alias{map_gatherings}
\title{Map Gathering Abbreviations and Full Names}
\usage{
map_gatherings(x, from = "Standard", to = "Name")
}
\arguments{
\item{x}{Gatherings vector}

\item{from}{Input terms to map (see \code{names(gatherings_table())} for options)}

\item{to}{Output terms to map (see \code{names(gatherings_table())} for options)}
}
\value{
Mapped gatherings.
}
\description{
Convert gatherings names between different formats.
}
\examples{
map_gatherings(c("8vo", "16mo"), from = "short", to = "long") 
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
