% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mark_dissertations.R
\name{mark_dissertations}
\alias{mark_dissertations}
\title{mark_dissertations}
\usage{
mark_dissertations(df)
}
\arguments{
\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Establish two binary variables: dissertation and synodal dissertation
}
\examples{
\dontrun{df <- mark_dissertations(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

