% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pagecounts.R
\name{mean_pagecounts}
\alias{mean_pagecounts}
\title{Mean Page Counts}
\usage{
mean_pagecounts(df)
}
\arguments{
\item{df}{data frame, including the required fields.}
}
\value{
Average page count.
}
\description{
Estimate page counts.
}
\examples{
\dontrun{m <- mean_pagecounts(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

