% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pagecounts_issue.R
\name{mean_pagecounts_issue}
\alias{mean_pagecounts_issue}
\title{mean_pagecounts_issue}
\usage{
mean_pagecounts_issue(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Average page count information
}
\description{
Estimate page counts for issues
}
\examples{
# m <- mean_pagecounts_issue(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

