% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pagecounts_multivol.R
\name{mean_pagecounts_multivol}
\alias{mean_pagecounts_multivol}
\title{Mean pagecounts multivolume}
\usage{
mean_pagecounts_multivol(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Average page count information
}
\description{
Calculate mean page counts for multi-volume documents.
}
\examples{
# m <- mean_pagecounts_multivol(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

