% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pagecounts_univol.R
\name{mean_pagecounts_univol}
\alias{mean_pagecounts_univol}
\title{Mean pagecounts for single volume documents}
\usage{
mean_pagecounts_univol(df)
}
\arguments{
\item{df}{data frame}
}
\value{
Average page count information
}
\description{
Calculate mean page counts for single volume documents.
}
\examples{
# m <- mean_pagecounts_univol(df)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

