% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_initials.R
\name{name_initials}
\alias{name_initials}
\title{name_initials}
\usage{
name_initials(x, format = NULL)
}
\arguments{
\item{x}{a vector of names}

\item{format}{Name format. By default interpreted as First Second Last, with the corresponding initials FSL.}
}
\value{
A vector of initials
}
\description{
Pick name initials
}
\examples{
name_initials("Thomas Albert")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

