% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names.R
\name{notnames}
\alias{notnames}
\title{notnames}
\usage{
notnames(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Vector of last names
}
\description{
List of strings that are not names (manually checked)
}
\examples{
\dontrun{x <- notnames()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

