% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{order_gatherings}
\alias{order_gatherings}
\title{order_gatherings}
\usage{
order_gatherings(x)
}
\arguments{
\item{x}{A vector or factor of gathering data}
}
\value{
A factor with ordered gatherings levels
}
\description{
Sort gatherings levels
}
\examples{
# order_gatherings(factor(c("2to", "1to", "8to")))
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

