% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{parse_collation_formula}
\alias{parse_collation_formula}
\title{Parse collation formula}
\usage{
parse_collation_formula(x)
}
\arguments{
\item{x}{input char}
}
\value{
vector containing numbers of leaves
}
\description{
internal
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{internal}
