% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_firstname.R
\name{pick_firstname}
\alias{pick_firstname}
\title{pick_firstname}
\usage{
pick_firstname(x, format = "last, first")
}
\arguments{
\item{x}{a vector of full names}

\item{format}{name format}
}
\value{
a vector of first names
}
\description{
Pick first name from full name, assuming the format is known
}
\examples{
pick_firstname("Hobbes, Thomas")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

