% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_lastname.R
\name{pick_lastname}
\alias{pick_lastname}
\title{pick_lastname}
\usage{
pick_lastname(x, format = "last, first")
}
\arguments{
\item{x}{a vector of full names}

\item{format}{name format}
}
\value{
a vector of last names
}
\description{
Pick last name from full name, assuming the format is known
}
\examples{
pick_lastname("Hobbes, Thomas")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

