% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/volume.R
\name{pick_multivolume}
\alias{pick_multivolume}
\title{pick_multivolume}
\usage{
pick_multivolume(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Volume information
}
\description{
Pick volume information for multi-volumen documents
}
\examples{
pick_multivolume("v.1-3, 293")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

