% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_parts.R
\name{pick_parts}
\alias{pick_parts}
\title{Pick Parts}
\usage{
pick_parts(x)
}
\arguments{
\item{x}{physical_extent field. Vector or factor of strings.}
}
\value{
Volume information
}
\description{
Pick parts information.
}
\examples{
\dontrun{pick_parts("2 parts")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
