% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publisher_forby.R
\name{pick_print_fields}
\alias{pick_print_fields}
\title{pick_print_fields}
\usage{
pick_print_fields(x, field)
}
\arguments{
\item{x}{A vector}

\item{field}{field names}
}
\value{
fields
}
\description{
Pick print fields
}
\examples{
\dontrun{x2 <- pick_print_fields(x, field)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

