% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_volume.R
\name{pick_volume}
\alias{pick_volume}
\title{Pick volume}
\usage{
pick_volume(s)
}
\arguments{
\item{s}{Page number field. Vector or factor of strings.}
}
\value{
Volume
}
\description{
Pick volume
}
\details{
A single document, but check which volume
}
\examples{
\dontrun{pick_volume("v.4")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
