% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/place2country.R
\name{place2country}
\alias{place2country}
\title{Region-country mapping table}
\usage{
place2country(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
data.frame
}
\description{
Region-country mapping table
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

