% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Timeline Plot}
\usage{
plot_timeline(x, y = NULL, field = "titlecount", group = NULL, nmin = 0,
  mode = "absolute", time.window = 10)
}
\arguments{
\item{x}{First data.frame}

\item{y}{Second data.frame. Optional, for comparison.}

\item{field}{Numeric field to summarize in the timeline. The number of entries (title count) per decade is shown by default, But if this argument is used, the sum of entries per decade for this field is given.}

\item{group}{Optional. Name for a data field that indicates groups to compare. If given, the different groups are indicated by lines.}

\item{nmin}{Include only entries with at least nmin absolute frequency}

\item{mode}{"absolute" or "relative"}

\item{time.window}{Time window for the timeline in years. Default: 10 (publication decade).}
}
\value{
List:
\itemize{
  \item{plot}{ggplot object}
  \item{table}{summary table}
}
}
\description{
Plot timeline for selected variable (possibly across various groups).
}
\details{
If two data.frames (x,y) are provided in the function arguments, a comparison plot will be generated.
}
\examples{
\dontrun{timeline(df, "gatherings")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

