% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_all.R
\name{polish_all}
\alias{polish_all}
\title{Generic Preprocessing for Bibliographic Metadata}
\usage{
polish_all(df.orig, fields = NULL, verbose = TRUE, file = NULL,
  mc.cores = 1, conversions = list(), languages = NULL)
}
\arguments{
\item{df.orig}{Original data.frame}

\item{fields}{Fields to be preprocessed.}

\item{verbose}{verbose}

\item{file}{Temporary results saved into a file}

\item{mc.cores}{Number of cores for parallelization}

\item{conversions}{Field conversion list}

\item{languages}{Languages to be used in polishing}
}
\value{
Preprocessed data.frame and field conversion list.
}
\description{
Polish all fields.
}
\examples{
\dontrun{a <- polish_all(df.orig)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

