% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_author.R
\name{polish_author}
\alias{polish_author}
\title{polish_author}
\usage{
polish_author(s, stopwords = NULL)
}
\arguments{
\item{s}{Vector of author names}

\item{stopwords}{Stopwords}
}
\value{
Polished vector
}
\description{
Polish author
}
\examples{
s2 <- polish_author("Smith, William")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

