% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_corporate.R
\name{polish_corporate}
\alias{polish_corporate}
\title{Polish corporate}
\usage{
polish_corporate(x)
}
\arguments{
\item{x}{A vector of corporate entries}
}
\value{
Polished vector
}
\description{
Polish corporate
}
\examples{
# x2 <- polish_corporate("unknown")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
