% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{polish_dimension}
\alias{polish_dimension}
\title{polish_dimension}
\usage{
polish_dimension(s)
}
\arguments{
\item{s}{A dimension note (a single string of one document)}

\item{sheetsizes}{Sheet size conversion table}
}
\value{
Polished dimension information with the original string
	   and gatherings and cm information collected from it
}
\description{
Polish dimension field of a single document
}
\examples{
# polish_dimension("4")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{internal}

