% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{polish_dimensions}
\alias{polish_dimensions}
\title{polish_dimensions}
\usage{
polish_dimensions(s)
}
\arguments{
\item{s}{A vector of dimension notes}
}
\value{
Table
}
\description{
Polish dimension field for many documents at once
}
\examples{
# polish_dimensions(c("2", "4"))
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{utilities}

