% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_gatherings.R
\name{polish_gatherings}
\alias{polish_gatherings}
\title{polish_gatherings}
\usage{
polish_gatherings(x)
}
\arguments{
\item{x}{A vector or factor of gathering data}
}
\value{
A factor with polished gatherings levels
}
\description{
Polish gatherings
}
\examples{
polish_gatherings(factor(c("2fo", "1tolded", "8vo")))
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
