% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_lifespan_of_author.R
\name{polish_lifespan_of_author}
\alias{polish_lifespan_of_author}
\title{polish_lifespan_of_author}
\usage{
polish_lifespan_of_author(df)
}
\arguments{
\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Pick and polish years of birth and death
}
\examples{
\dontrun{df <- polish_lifespan_of_author(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

