% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polish_page}
\alias{polish_page}
\title{polish_page}
\usage{
polish_page(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Cleaned up version of page number field.
}
\description{
clean up page numbers
}
\examples{
# polish_page("4p.")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{internal}

