% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polish_pages}
\alias{polish_pages}
\title{polish_pages}
\usage{
polish_pages(s, verbose = FALSE)
}
\arguments{
\item{s}{Page number field. Vector or factor of strings.}

\item{verbose}{Print progress info}
}
\value{
Raw and estimated pages per document part
}
\description{
clean up page numbers
}
\details{
Document parts are separated by semicolons
}
\examples{
# polish_pages("4p.")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

