% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_physical_extent.R
\name{polish_physext_help}
\alias{polish_physext_help}
\title{Polish physical_extent help field}
\usage{
polish_physext_help(s, verbose, page.synonyms, page.harmonize, sheet.harmonize)
}
\arguments{
\item{s}{Input char}

\item{verbose}{Print progress info}
}
\value{
Internal
}
\description{
Internal
}
\examples{
# TBA
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{?}
\keyword{A}
\keyword{but}
\keyword{check}
\keyword{document,}
\keyword{internal}
\keyword{number}
\keyword{single}
\keyword{volume}
\keyword{which}

