% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_physical_extent.R
\name{polish_physext_help2}
\alias{polish_physext_help2}
\title{Polish physical_extent help field 2}
\usage{
polish_physext_help2(s, page.synonyms, page.harmonize, sheet.harmonize, harm.pi)
}
\arguments{
\item{s}{Input char}

\item{page.synonyms}{page.synonyms}

\item{vols}{vols}
}
\value{
Internal
}
\description{
Internal
}
\examples{
# TBA
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{internal}

