% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_physical_extent.R
\name{polish_physical_extent}
\alias{polish_physical_extent}
\title{Polish physical_extent Field}
\usage{
polish_physical_extent(x, verbose = FALSE, rm.dim.file = NULL)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}

\item{verbose}{Print progress info}
}
\value{
Raw and estimated pages per document part
}
\description{
Pick page counts, volume counts and volume numbers.
}
\details{
Document parts are separated by semicolons

A summary of page counting rules that this function aims to (approximately) implement are provided in 
\url{https://www.libraries.psu.edu/psul/cataloging/training/bpcr/300.html}
}
\examples{
tab <- polish_physical_extent("4p.", verbose = TRUE)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
