% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonize_place.R
\name{polish_place}
\alias{polish_place}
\title{polish_place}
\usage{
polish_place(x)
}
\arguments{
\item{x}{A vector of place names}
}
\value{
Polished vector
}
\description{
Polish place
}
\examples{
\dontrun{x2 <- polish_place(x, file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

