% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publication_frequency.R
\name{polish_publication_frequency}
\alias{polish_publication_frequency}
\title{Polish Publication Frequency}
\usage{
polish_publication_frequency(x)
}
\arguments{
\item{x}{publication frequency field (a vector)}
}
\description{
Harmonize publication frequencies.
}
\examples{
\dontrun{df <- polish_publication_frequency("Kerran vuodessa")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
