% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publication_frequency.R
\name{polish_publication_frequency_english}
\alias{polish_publication_frequency_english}
\title{Polish Publication Frequency English}
\usage{
polish_publication_frequency_english(x)
}
\arguments{
\item{x}{publication frequency field (a vector)}
}
\description{
Harmonize publication frequencies for English data.
}
\examples{
\dontrun{df <- polish_publication_frequency_english("weekly")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
