% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publication_frequency.R
\name{polish_publication_frequency_finnish}
\alias{polish_publication_frequency_finnish}
\title{Polish Publication Frequency Finnish}
\usage{
polish_publication_frequency_finnish(x)
}
\arguments{
\item{x}{publication frequency field (a vector) 1}
}
\description{
Harmonize publication frequencies for Finnish data.
}
\examples{
\dontrun{df <- polish_publication_frequency_finnish("Kerran vuodessa")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
