% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publication_frequency.R
\name{polish_publication_frequency_swedish}
\alias{polish_publication_frequency_swedish}
\title{Polish Publication Frequency Swedish}
\usage{
polish_publication_frequency_swedish(x)
}
\arguments{
\item{x}{publication frequency field (a vector)}
}
\description{
Harmonize publication frequencies for Swedish data.
}
\examples{
\dontrun{df <- polish_publication_frequency_swedish("1 nr/ar")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

