% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{polish_publisher}
\usage{
polish_publisher(df)
}
\arguments{
\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Polish publishing house names
}
\examples{
\dontrun{df <- polish_publisher(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

