% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publisher_forby.R
\name{polish_publisher_forby}
\alias{polish_publisher_forby}
\title{Polish publisher}
\usage{
polish_publisher_forby(x)
}
\arguments{
\item{x}{Publisher vector}
}
\value{
Polished vector
}
\description{
Polish publisher field separating for/by
}
\details{
Polish publisher field.
}
\examples{
# polish_publisher("printed and sold by R. Marchbank")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

