% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_physical_extent.R
\name{polish_signature_statement_pagecount}
\alias{polish_signature_statement_pagecount}
\title{Polish signature statements}
\usage{
polish_signature_statement_pagecount(s)
}
\arguments{
\item{s}{input char}
}
\value{
vector
}
\description{
Internal
}
\references{
See citation("bibliographica") and explanations of signature statements in \url{https://collation.folger.edu/2016/05/signature-statements/} and \url{https://manual.stcv.be/p/Inputting_Collation_Data_in_Brocade}.
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}
