% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_university.R
\name{polish_university}
\alias{polish_university}
\title{polish_university}
\usage{
polish_university(df)
}
\arguments{
\item{df}{Main dataframe}
}
\value{
Main dataframe
}
\description{
Polish university names
}
\examples{
\dontrun{df <- polish_university(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

