% Generated by roxygen2 (4.0.2): do not edit by hand
\name{polish_volumecount}
\alias{polish_volumecount}
\title{polish_volumecount}
\usage{
polish_volumecount(s)
}
\arguments{
\item{s}{Page number field. Vector or factor of strings.}
}
\value{
Number of volumes
}
\description{
Get volume number from page field if available
}
\details{
Refers to multi-volume document where the number of volumes has been specified
}
\examples{
polish_volumecount("4v.")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

