% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_year_of_publication.R
\name{polish_year_of_publication}
\alias{polish_year_of_publication}
\title{polish_year_of_publication}
\usage{
polish_year_of_publication(df)
}
\arguments{
\item{df}{Dataframe with modified fields}
}
\value{
Dataframe with modified fields
}
\description{
Pick and polish year(s) of publication
}
\examples{
\dontrun{df <- polish_year_of_publication(df)}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

