% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_combined_catalogs.R
\name{prepare_combined_catalogs}
\alias{prepare_combined_catalogs}
\title{Prepare combined catalogs}
\usage{
prepare_combined_catalogs(df.fennica, df.kungliga, first_word_count = 5)
}
\arguments{
\item{df.fennica}{Raw Fennica data}

\item{df.kungliga}{Raw Kungliga data}

\item{first_word_count}{Number of words in the beginning of a title needed to make a match}
}
\value{
Data frame of core fields
}
\description{
Preprocesses and combines two catalogs (not every field)
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

