% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_frequency_text.R
\name{publication_frequency_text}
\alias{publication_frequency_text}
\title{Annual Publication Frequency to Text}
\usage{
publication_frequency_text(x, peryear)
}
\arguments{
\item{x}{publication frequency per year}
}
\description{
Convert annual publication frequencies to text format.
}
\examples{
\dontrun{df <- publication_frequency_text("Once a year", 1)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

