% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mapping.R
\name{read_mapping}
\alias{read_mapping}
\title{Read Mapping}
\usage{
read_mapping(file, mode = "table", sep = ";", self.match = FALSE,
  include.lowercase = FALSE, ignore.empty = FALSE, sort = FALSE,
  verbose = FALSE, remove.ambiguous = TRUE, lowercase = FALSE,
  from = "synonyme", to = "name", fast = FALSE, encoding = "UTF-8")
}
\arguments{
\item{file}{Input file}

\item{mode}{The input file type: "list" or "table"; see details.}

\item{sep}{Separator mark.}

\item{self.match}{Include self matches}

\item{include.lowercase}{Include lowercase}

\item{ignore.empty}{Ignore the empty cases}

\item{sort}{Sort synonymes}

\item{verbose}{verbose}

\item{remove.ambiguous}{Remove ambiguous terms.}

\item{lowercase}{All synonymes considered in lowercase only}

\item{from}{field that will be replaced}

\item{to}{field that contains the final names}

\item{fast}{Use the fast method fread (sensitive to problems in table format)}

\item{encoding}{Optional encoding of the input file (see help(readLines))}
}
\value{
Synonyme data frame with the fields 'name' (the selected term) and 'synonyme' (the alternative terms).
}
\description{
Read mapping table.
}
\details{
If mode = "list", each row of the input file corresponds to a unique entry with potentially multiple name variants, separated by semicolon. The first element gives the selected version of the name, the subsequent elements list synonymes that will be mapped to the selected version. If mode = "table", the file has two columns where each row corresponds to a unique entry and has the selected name and a single alternative name.
}
\examples{
\dontrun{syn <- read_mapping(file)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

