% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parsed_field.R
\name{read_parsed_fields}
\alias{read_parsed_fields}
\title{Load Parsed Field}
\usage{
read_parsed_fields(f, field = 300, subfield = "c", n = NULL)
}
\arguments{
\item{f}{Datafile in csv}

\item{field}{Field to read}

\item{subfield}{Subfield to read}

\item{n}{n}
}
\value{
data.frame with raw data fields
}
\description{
Load parsed CSV fields.
}
\examples{
\dontrun{df <- read_parsed_fields(f)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti
}
\keyword{utilities}
