% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tables.R
\name{region2country}
\alias{region2country}
\title{Region-country mapping table}
\usage{
region2country(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
data.frame
}
\description{
Region-country mapping table
}
\examples{
map <- region2country()
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

