% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_brackets_from_letters.R
\name{remove_brackets_from_letters}
\alias{remove_brackets_from_letters}
\title{Remove Brackets from Letters}
\usage{
remove_brackets_from_letters(x, myletters = NULL)
}
\arguments{
\item{x}{A vector}

\item{myletters}{Letters to remove}
}
\value{
A polished vector
}
\description{
Remove brackets surrounding letters.
}
\examples{
x2 <- remove_brackets_from_letters("[p]")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
