% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dimension.R
\name{remove_dimension}
\alias{remove_dimension}
\title{Remove Dimension Data}
\usage{
remove_dimension(x, terms)
}
\arguments{
\item{x}{A character vector that may contain dimension information}
}
\value{
The character vector with dimension information removed
}
\description{
Remove dimension data.
}
\examples{
# remove_dimension("4to 40cm", sheet_sizes())
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}
