% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{remove_dimension_info}
\alias{remove_dimension_info}
\title{remove_dimension_info}
\usage{
remove_dimension_info(x, sheetsizes)
}
\arguments{
\item{x}{A character vector that may contain dimension information}
}
\value{
The character vector with dimension information removed
}
\description{
Remove dimension information from a single document
}
\examples{
# s <- remove_dimension("4to", sheet_sizes())
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("estc")
}
\keyword{internal}

