% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicates}
\usage{
remove_duplicates(df.preprocessed, field_names = c("short_title",
  "publication_place", "publication_year"))
}
\arguments{
\item{field_names}{List of field names of exactly matching values}

\item{df.prepocessed}{Data frame of combined catalogs}
}
\value{
Data frame
}
\description{
Tags files as to be removed
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

