% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_endings}
\alias{remove_endings}
\title{remove_endings}
\usage{
remove_endings(x, endings)
}
\arguments{
\item{x}{vector}

\item{endings}{endings to remove}
}
\value{
polished vector
}
\description{
Remove specified endings of strings
}
\examples{
\dontrun{x2 <- remove_endings(x, endings)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

