% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_letters.R
\name{remove_letters}
\alias{remove_letters}
\title{remove_letters}
\usage{
remove_letters(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Polished vector
}
\description{
Remove specific single letters
}
\examples{
\dontrun{remove_letters(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

