% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_print_statements.R
\name{remove_print_statements}
\alias{remove_print_statements}
\title{remove_print_statements}
\usage{
remove_print_statements(x, remove.letters = FALSE, n.iter = 1)
}
\arguments{
\item{x}{a vector}

\item{remove.letters}{Remove individual letters TRUE/FALSE}

\item{n.iter}{Number of iterative repetitions of this function}
}
\value{
Polished vector
}
\description{
Remove print statements
}
\examples{
x2 <- remove_print_statements("Printed in London")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

