% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_sl.R
\name{remove_sl}
\alias{remove_sl}
\title{Remove s.l etc. clauses}
\usage{
remove_sl(x, terms = NULL)
}
\arguments{
\item{x}{A character vector}

\item{terms}{Terms to be removed (a character vector). Optional}
}
\value{
Polished vector
}
\description{
Remove s.l etc. clauses
}
\examples{
x2 <- remove_sl(c("s.l.", "London"))
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
