% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_stopwords.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{Remove Stopwords}
\usage{
remove_stopwords(x, terms = NULL)
}
\arguments{
\item{x}{A vector}

\item{terms}{Stopwords}
}
\value{
A vector polished
}
\description{
Remove stopwords from input vector.
}
\examples{
# x2 <- remove_stopwords(c("a", "well", "james"), "well")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

