% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_terms_all.R
\name{remove_terms_all}
\alias{remove_terms_all}
\title{Remove Terms Completely}
\usage{
remove_terms_all(
  x,
  terms,
  include.lowercase = FALSE,
  polish = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{x}{A vector}

\item{terms}{Terms to be removed}

\item{include.lowercase}{Include also lowercase versions of the terms}

\item{polish}{polish the entries after removing the terms (remove trailing spaces and periods)}

\item{recursive}{Apply the changes recursively along the list ?}
}
\value{
Vector with terms removed
}
\description{
Fully Remove the given terms from the strings.
}
\details{
After removing the numerics, beginning, double and ending 
         spaces are also removed from the strings.
}
\examples{
\dontrun{x2 <- remove_terms_all(x, terms)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
