% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_time_info.R
\name{remove_time_info}
\alias{remove_time_info}
\title{remove_time_info}
\usage{
remove_time_info(x)
}
\arguments{
\item{x}{Vector (time field)}
}
\value{
Polished vector
}
\description{
Remove time information
}
\details{
Remove months, year terms and numerics
}
\examples{
\dontrun{x2 <- remove_time_info(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

