% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/self_published.R
\name{self_published}
\alias{self_published}
\title{Identify Self published documents}
\usage{
self_published(df)
}
\arguments{
\item{df}{data.frame that includes the given field}
}
\value{
Output of the polished field
}
\description{
Identify Self published documents
}
\details{
The function considers self published documents those that have 'author' (or similar) in the publisher field, and those where the publisher is unknown.
}
\examples{
\dontrun{a <- polish_field(df, "title")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

