% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sheet_sizes.R
\name{sheet_sizes}
\alias{sheet_sizes}
\title{Sheet Sizes}
\usage{
sheet_sizes(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Sheet size table
}
\description{
Read sheet size table.
}
\examples{
sheetsizes <- sheet_sizes()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
