% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitpick.R
\name{splitpick}
\alias{splitpick}
\title{Pick Substring Indicated by Separator}
\usage{
splitpick(x, sep, which)
}
\arguments{
\item{x}{Input string}

\item{sep}{Separator string}

\item{which}{Indicate which part to pick}
}
\value{
Polished string
}
\description{
Split the string by separator and pick the former or latter part.
}
\examples{
x2 <- splitpick("London re Edinburgh", " re ", 1)
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
