% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_comhis}
\alias{theme_comhis}
\title{Theme function for COMHIS group}
\usage{
theme_comhis(type = "invalid", base_size = 12)
}
\arguments{
\item{type}{Type of the data to be plotted ("continuous"; "discrete")}
}
\description{
default theme for consistent plot style within the 
COMHIS group.
}
\examples{
  #library(ggplot2)
  #data(iris)
  #p <- ggplot(iris, aes(x = Sepal.Width, y = Sepal.Length)) +
  #  geom_point()
  #p + theme_comhis("continuous")
}
\author{
Abhishekh Gupta \email{abishak.gupta@gmail.com}
}
