% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_relative.R
\name{timeline_relative}
\alias{timeline_relative}
\title{Retrieve Relative Timeline}
\usage{
timeline_relative(x, y, myfield, time.window = 10)
}
\arguments{
\item{x}{First data frame}

\item{y}{Second data frame}

\item{myfield}{Numeric field to summarize in the timeline (x/y). The number of entries (title count) per decade is used by default. If this argument is used, the sum of entries per decade for this field is given.}

\item{time.window}{Time window for the timeline in years. Default: 10 (publication decade).}
}
\value{
data.frame
}
\description{
Relative timeline for selected variable between two data sets.
}
\examples{
\dontrun{timeline_relative(df, "gatherings")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
