% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_names.R
\name{trim_names}
\alias{trim_names}
\title{trim_names}
\usage{
trim_names(s, stopwords, remove.letters = FALSE)
}
\arguments{
\item{s}{A vector of names}

\item{stopwords}{stopwords}

\item{remove.letters}{Remove individual letters}
}
\value{
Polished vector
}
\description{
Trim names
}
\examples{
s2 <- trim_names("james", "and")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

