% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trimming}
\alias{trimming}
\title{trimming}
\usage{
trimming(x, n = 1)
}
\arguments{
\item{x}{vector}

\item{n}{iterations}
}
\value{
Polished vector
}
\description{
Trim strings
}
\examples{
\dontrun{x2 <- trimming(x, n = 1)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

