% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{unfold_set}
\alias{unfold_set}
\title{Unfold a gathering set}
\usage{
unfold_set(start, end, n1, n2, sig_table)
}
\arguments{
\item{start}{the signature of first gathering in the set}

\item{end}{the signature of the last gathering (or NA if non-sequence)}

\item{n1}{the number of leaves per gathering}

\item{n2}{the second number of leaves in case of alternating numbers (or NA
if not present)}

\item{sig_table}{the table of admissible gathering signatures and their
sequential ordering (see \code{\link{build_signature_table}})}
}
\value{
a data frame giving the signature and number of leaves of each
        gathering
}
\description{
internal
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{internal}
