% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_collation_formula.R
\name{unfold_sets}
\alias{unfold_sets}
\title{Unfold gathering sets}
\usage{
unfold_sets(x)
}
\arguments{
\item{x}{the list of dataframes containing recognized gathering sets (output
of \code{\link{parse_collation_formula}})}
}
\value{
a data frame giving the signatures and numbers of leaves of each
        gathering in each formula
}
\description{
This is a wrapper around \code{\link{unfold_set}} operating on
             data frames resulting from parsing multiple formulas.
}
\author{
Maciej Janicki \email{maciej.janicki@helsinki.fi}
}
\keyword{internal}
