% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{validate_names}
\alias{validate_names}
\title{validate_names}
\usage{
validate_names(namelist, database)
}
\arguments{
\item{namelist}{Vector of names to be validated}

\item{database}{Specify the name database to be used}
}
\value{
List with following elements:
   \itemize{
     \item validated Logical vector indicating the valid names
\item invalid List of invalid name components
   }
}
\description{
Validate names
}
\examples{
\dontrun{x <- validate_names(namelist, database)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

