% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_preprocessed_data}
\alias{validate_preprocessed_data}
\title{Validate Preprocessed Data}
\usage{
validate_preprocessed_data(data.preprocessed, max.pagecount = 5000)
}
\arguments{
\item{data.preprocessed}{Preprocessed data.}

\item{max.pagecount}{Upper gap for the pagecount for ocs that exceed this limit.}
}
\value{
Modified data.
}
\description{
Preprocessing validators and some adjustments.
}
\examples{
# \dontrun{validate_preprocessed_data(data.preprocessed)}
}
\references{
See citation("bibliographica")
}
\author{
Ville Vaara and Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
