% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_publication_time.R
\name{validate_publication_time}
\alias{validate_publication_time}
\title{Validate Publication Time}
\usage{
validate_publication_time(data.preprocessed, min.year = NULL,
  max.year = NULL)
}
\arguments{
\item{data.preprocessed}{Preprocessed data.}

\item{min.year}{Minimum accepted year; all accepted by default}

\item{max.year}{Maximum accepted year; current year by default}
}
\value{
Modified data.
}
\description{
Preprocessing validator.
}
\examples{
# \dontrun{validate_publication_time(data.preprocessed)}
}
\references{
See citation("bibliographica")
}
\author{
Ville Vaara and Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
