% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_towns_in_catalog.R
\name{write_towns_in_catalog}
\alias{write_towns_in_catalog}
\title{Write towns in catalog}
\usage{
write_towns_in_catalog(combined_catalogs, catalog_name, towns,
  write_to_one_file = FALSE)
}
\arguments{
\item{combined_catalogs}{Data frame of combined catalogs}

\item{catalog_name}{String of the catalog whose files are listed}

\item{towns}{Names of towns that will be listed}

\item{write_to_file}{Boolean indicating if the value should be printed or not}
}
\value{
Boolean Returns always TRUE
}
\description{
Lists & writes publications printed in specific towns from a given catalog
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

