% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ambiguous_authors_table.R
\name{ambiguous_authors_table}
\alias{ambiguous_authors_table}
\title{Ambiguous Authors Table}
\usage{
ambiguous_authors_table(file = NULL)
}
\arguments{
\item{file}{Input file}
}
\value{
Author synonyme data frame with fields 'name' and 'synonyme'
}
\description{
Read table of ambiguous author name synonymes
}
\details{
Each row of the input file corresponds to a unique author with potentially multiple name variants, separated by semicolon. The first column gives the accepted version of the name, the other columns list synonymes that will be mapped to the accepted version.
}
\examples{
aa <- ambiguous_authors_table()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
