% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_years.R
\name{approximate_years}
\alias{approximate_years}
\title{Approximate Years}
\usage{
approximate_years(x, polished_year = NULL)
}
\arguments{
\item{x}{year field (a vector)}

\item{polished_year}{Polished year information}
}
\value{
data.frame with the fields 'start' and 'end'
}
\description{
Approximate year for entries that are approximations.
}
\examples{
\dontrun{df <- polish_years(c("1746", "1745-1750"))}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti and Hege Roivainen email{leo.lahti@iki.fi}
}
\keyword{utilities}
