% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment_dimension_table.R
\name{augment_dimension_table}
\alias{augment_dimension_table}
\title{augment_dimension_table}
\usage{
augment_dimension_table(dimension.table, dimtab, verbose = FALSE,
  sheet.dimension.table)
}
\arguments{
\item{dimension.table}{dimension.table}

\item{dimtab}{Mapping between document dimensions.}

\item{verbose}{verbose}
}
\value{
Augmented dimension.table
}
\description{
Estimate missing entries in dimension table where possible
}
\examples{
# augment_dimension_table(dimension.table)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\seealso{
polish_dimensions
}
\keyword{utilities}

