% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/author_unique.R
\name{author_unique}
\alias{author_unique}
\title{Create unique author ID}
\usage{
author_unique(df, format = "last, first", initialize.first = FALSE)
}
\arguments{
\item{df}{data.frame with fields "author_name", "author_birth", "author_death"}

\item{format}{name format ("last, first" or "first last")}

\item{initialize.first}{Convert first names into initials. Useful for removing duplicates when the name list contains different versions with both full name and the initials. Use of the life year fields helps to avoid mixing identical abbreviated names.}
}
\value{
Character vector of unique author IDs
}
\description{
Form unique author identifiers by combining name and life years
}
\examples{
\dontrun{a <- author_unique(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
