% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bibliographica-package.R
\docType{package}
\name{bibliographica-package}
\alias{bibliographica}
\alias{bibliographica-package}
\title{bibliographica}
\description{
Brief summary of the bibliographica package
}
\details{
\tabular{ll}{
Package: \tab bibliographica\cr
Type: \tab Package\cr
Version: \tab See sessionInfo() or DESCRIPTION file\cr
Date: \tab 2014-2015\cr
License: \tab BSD-2-clause\cr
LazyLoad: \tab yes\cr
}
}
\examples{
library(bibliographica)
}
\author{
Leo Lahti et al. \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
\url{https://github.com/ropengov/bibliographica}
}
\keyword{package}

