% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capitalize.R
\name{capitalize}
\alias{capitalize}
\title{Capitalize strings}
\usage{
capitalize(x, format = "first.letter")
}
\arguments{
\item{x}{a character vector to capitalize}

\item{format}{Capitalization format: "first.letter" (first letter
capitalized) or "all.words" (all words capitalized)}
}
\value{
Capitalized character vector
}
\description{
Capitalize the first letter.
}
\examples{
capitalize("print")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
