% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_volumes.R
\name{check_volumes}
\alias{check_volumes}
\title{Check Volumes}
\usage{
check_volumes(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Volume information
}
\description{
Pick volume information from page number field
}
\examples{
\dontrun{check_volumes("v.4, 293")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
