% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_publisher.R
\name{clean_publisher}
\alias{clean_publisher}
\title{Clean Publisher}
\usage{
clean_publisher(x, languages = c("english"))
}
\arguments{
\item{x}{A vector of publisher names}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Vector of publisher names
}
\description{
Preliminary cleanup for the publisher field.
}
\examples{
# clean_publisher(x, languages=c("finnish", "swedish", "latin"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
