% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compress_field.R
\name{compress_field}
\alias{compress_field}
\alias{collapse}
\title{Compress Field into Fewer Categories}
\usage{
compress_field(x, topn = NULL, min.freq = NULL, rest = "Other")
}
\arguments{
\item{x}{Vector or factor.}

\item{topn}{Number of top entries to include; or a character vector}

\item{min.freq}{Minimum frequency for the included groups.}

\item{rest}{Name to be used for the remaining entries}
}
\value{
Compressed version of the input.
}
\description{
Merge the less common terms into a single group
}
\details{
With the min.freq argument, groups that have less items
         that this will be merged.
}
\examples{
  p <- compress_field(c("A", "A", "A", "B","B","C", "D"),
         topn = 2, rest = "Other")
  p <- compress_field(c("A", "A", "A", "B","B","C", "D"),
         topn = c("A", "B"), rest = "Other")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
