% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condense_spaces.R
\name{condense_spaces}
\alias{condense_spaces}
\title{Condense Spaces}
\usage{
condense_spaces(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector with extra spaces removed
}
\description{
Trim and remove double spaces from the input strings.
}
\details{
Beginning, double and ending spaces are also removed from the strings.
}
\examples{
x2 <- condense_spaces(" a  b cd ") # "a b cd"
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
