% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_synonymes.R
\name{dimension_synonymes}
\alias{dimension_synonymes}
\title{Return dimension synonyme table}
\usage{
dimension_synonymes(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Dimension synonyme table
}
\description{
Return dimension synonyme table
}
\examples{
dimension_synonymes()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
