% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_table.R
\name{dimension_table}
\alias{dimension_table}
\title{Dimension Table}
\usage{
dimension_table(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Document dimension table
}
\description{
Document dimension mapping table.
}
\examples{
dimension_table()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
