% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_author.R
\name{enrich_author}
\alias{enrich_author}
\title{Enrich Author Field}
\usage{
enrich_author(df)
}
\arguments{
\item{df}{Preprocessed data.frame}
}
\value{
Augmented data.frame
}
\description{
Enrich author field.
}
\examples{
\dontrun{df2 <- enrich_authoragecount(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
