% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_pagecount.R
\name{enrich_pagecount}
\alias{enrich_pagecount}
\title{Enrich Page Count}
\usage{
enrich_pagecount(x, estimate = "median.pages")
}
\arguments{
\item{x}{Preprocessed data.frame}

\item{estimate}{Which estimate to use for pagecount ("median.pages" / "mean.pages")}
}
\value{
Augmented data.frame
}
\description{
Augment missing pagecounts based on mean estimates from available data.
}
\examples{
\dontrun{df2 <- enrich_pagecount(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
