% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_preprocessed_data.R
\name{enrich_preprocessed_data}
\alias{enrich_preprocessed_data}
\title{Enrich Data}
\usage{
enrich_preprocessed_data(data.validated, df.orig)
}
\arguments{
\item{data.validated}{Validated data.frame}

\item{df.orig}{Original data.frame}
}
\value{
Augmented data.frame
}
\description{
Enrich data.
}
\examples{
\dontrun{df2 <- enrich_preprocessed_data(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
