% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/estimate_document_dimension.R
\name{estimate_document_dimensions}
\alias{estimate_document_dimensions}
\title{Estimate missing dimension information}
\usage{
estimate_document_dimensions(gatherings = NA, height = NA, width = NA,
  dimension.table = NULL, sheet.dimension.table = NULL)
}
\arguments{
\item{gatherings}{Available gatherings information}

\item{height}{Available height information}

\item{width}{Available width information}

\item{dimension.table}{Document dimension table (from dimension_table())}

\item{sheet.dimension.table}{Table to estimate sheet area.
If not given, the table given by sheet_sizes() is used by default.}
}
\value{
Augmented dimension information
}
\description{
Estimate missing dimension information
}
\examples{
# estimate_document_dimensions(gatherings = 2, height = 44)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

