% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pages_issue.R
\name{estimate_pages_issue}
\alias{estimate_pages_issue}
\title{Page estimation for issues}
\usage{
estimate_pages_issue(df, mean.pagecounts.issue)
}
\arguments{
\item{df}{data.frame of documents x variables}

\item{mean.pagecounts.issue}{Page count estimates to be used}
}
\value{
Page count estimates
}
\description{
Estimate pages for issues.
}
\examples{
\dontrun{p <- estimate_pages_issue(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

