% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pages_multivol.R
\name{estimate_pages_multivol}
\alias{estimate_pages_multivol}
\title{Page Estimation for Multivolume Documents}
\usage{
estimate_pages_multivol(df, pagecount.estimates)
}
\arguments{
\item{df}{data.frame of documents x variables}
}
\value{
Page count estimates
}
\description{
Estimate pages for multi-volume documents.
}
\examples{
\dontrun{p <- estimate_pages_multivol(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

