% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pages_singlevol.R
\name{estimate_pages_singlevol}
\alias{estimate_pages_singlevol}
\title{Page Estimation (Single Volume)}
\usage{
estimate_pages_singlevol(df, pagecount.estimates, estimate = "median.pages")
}
\arguments{
\item{df}{data.frame of documents x variables}

\item{pagecount.estimates}{Page count estimates to be used}

\item{estimate}{Which estimate to use for pagecount ("median.pages" / "mean.pages")}
}
\value{
Page count estimates
}
\description{
Estimate pages for single volume documents.
}
\examples{
\dontrun{p <- estimate_pages_singlevol(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
