% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_dimensions.R
\name{fill_dimensions}
\alias{fill_dimensions}
\title{Fill Missing Dimensions}
\usage{
fill_dimensions(x, dimension.table = NULL, sheet.dimension.table = NULL)
}
\arguments{
\item{x}{dimension string}

\item{dimension.table}{Given mappings between document dimensions}

\item{sheet.dimension.table}{Given mappings for sheet dimensions}
}
\value{
Augmented dimension vector
}
\description{
Estimate missing entries in dimension vector where possible.
}
\examples{
\dontrun{
  dimension.table <- dimension_table(); 
  fill_dimensions(c(original = NA, gatherings = NA, width = 30, height = 48), dimension.table)
}
}
\references{
See citation("bibliographica")
}
\seealso{
augment_dimension_table
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
