% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/names.R
\name{firstnames}
\alias{firstnames}
\title{firstnames}
\usage{
firstnames(...)
}
\arguments{
\item{...}{Arguments to be passed}
}
\value{
Vector of first names
}
\description{
First name list
}
\examples{
\dontrun{x <- firstnames()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

