% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstnames.R
\name{firstnames}
\alias{firstnames}
\title{Gender Table for First Names}
\usage{
firstnames(languages = NULL)
}
\arguments{
\item{languages}{Vector of language catalogues to include.}
}
\value{
Vector of first names
}
\description{
Combines first name - gender mappings from various sources in French, German, English, Finnish, and Custom lists, including pseudonymes.
}
\details{
Data sources:
  \itemize{
    \item{French}{Multilingual database \url{http://www.lexique.org/public/prenoms.php} (Prenoms.txt) and French \url{http://www.excel-downloads.com/forum/86934-liste-des-prenoms.htmlhttp://http://www.excel-downloads.com/forum/86934-liste-des-prenoms.html}.}
    \item{German}{\url{http://www.albertmartin.de/vornamen/}}
    \item{English}{For each year from 1880 to 2013, the number of children of each sex given each name. All names with more than 5 uses are given: \url{http://www.ssa.gov/oact/babynames/limits.html}). Here we use the implementation from the babynames R package. The name-gender mappings from different years are combined.}
    \item{Finnish}{Vaestorekisterikeskus VRK (see fennica::get_gender_fi) for details.}
    \item{Custom, Pseudonyme}{Provided by the authors of the bibliographica R package.}
  }
}
\examples{
\dontrun{x <- firstnames()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

