% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_summaryfiles.R
\name{generate_summaryfiles}
\alias{generate_summaryfiles}
\title{Generate Summary Pages}
\usage{
generate_summaryfiles(
  df,
  df.orig = NULL,
  author = "Author TBA",
  output.folder = "output.tables/",
  ntop = 20,
  summaries = c("overview", "author", "publicationplace", "publisher", "documents",
    "size", "gender", "topic", "language", "title", "publicationyear", "pagecount",
    "dimension")
)
}
\arguments{
\item{df}{data.frame to be summarized}

\item{df.orig}{Original unpolished data.frame}

\item{author}{Author name (to printed on the generated documents)}

\item{output.folder}{Path to the folder with associated summary table}

\item{ntop}{The number of top findings to show}

\item{summaries}{A character vector specifying the summaries to be produced.}
}
\value{
A vector of output file paths
}
\description{
Generates statistical summaries for preprocessed data.
}
\examples{
# generate_summaryfiles()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
