% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_country.R
\name{get_country}
\alias{get_country}
\title{Get Country}
\usage{
get_country(x, map = NULL)
}
\arguments{
\item{x}{A vector of region names (cities or municipalities etc.)}

\item{map}{data.frame with region to country mappings (fields 'region' and 'country')}
}
\value{
Country vector
}
\description{
Map geographic places to country names.
}
\examples{
x2 <- get_country("Berlin")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
