% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dissertation_mismatch.R
\name{get_dissertation_mismatch}
\alias{get_dissertation_mismatch}
\title{Get dissertation mismatches}
\usage{
get_dissertation_mismatch(
  df.preprocessed,
  field_names = c("short_title", "publication_place", "publication_year")
)
}
\arguments{
\item{df.preprocessed}{Data frame of combined catalogs}

\item{field_names}{List of field names of exactly matching values}
}
\value{
Data frame
}
\description{
Gets duplicates which are tagged as dissertations in another catalog, but in the other one aren't
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
