% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_gender}
\alias{get_gender}
\title{Get gender}
\usage{
get_gender(x, gendermap)
}
\arguments{
\item{x}{Vector of first names}

\item{gendermap}{Table with name-gender mappings}
}
\value{
Author gender information
}
\description{
Pick gender based on first names.
}
\examples{
  \dontrun{
    data(gendermap)
    get_gender(c("armi", "julius"), gendermap)
}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

