% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_gender}
\alias{get_gender}
\title{Get Gender}
\usage{
get_gender(x, gendermap)
}
\arguments{
\item{x}{Vector of first names}

\item{gendermap}{Table with name-gender mappings}
}
\value{
Author gender information
}
\description{
Pick gender based on first names.
}
\examples{
  \dontrun{
    gendermap <- read_mapping(system.file("extdata/gendermap.csv",
      package = "bibliographica"), sep = "\t", from = "name", to = "gender")
    get_gender(c("armi", "julius"), gendermap)
}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
