% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_relation_keyword.R
\name{get_relation_keyword}
\alias{get_relation_keyword}
\title{Get Relation Keyword}
\usage{
get_relation_keyword(x, full_name, languages = c("english"))
}
\arguments{
\item{x}{A character vector (e.g. publisher names)}

\item{full_name}{A vector of full names (from first name to last) it has been picked from publisher field}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Vector with relation keywords
}
\description{
Pick keyword denoting that the person in question is a relative to the named person
}
\examples{
# get_relation_keyword(x, full_name, languages=c("finnish", "swedish", "latin"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
