% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_at.R
\name{harmonize_at}
\alias{harmonize_at}
\title{harmonize_at}
\usage{
harmonize_at(x)
}
\arguments{
\item{x}{A vector}
}
\value{
Polished vector
}
\description{
Handle at statements
}
\examples{
\dontrun{x2 <- harmonize_at(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
