% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_dimension.R
\name{harmonize_dimension}
\alias{harmonize_dimension}
\title{Harmonize dimension}
\usage{
harmonize_dimension(x, synonyms = NULL)
}
\arguments{
\item{x}{A character vector that may contain dimension information}

\item{synonyms}{Synonyme table}
}
\value{
The character vector with dimension information harmonized
}
\description{
Harmonize dimension information
}
\examples{
\dontrun{harmonize_dimension("fol.")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}
