% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_field_names.R
\name{harmonize_field_names}
\alias{harmonize_field_names}
\title{Harmonize field names}
\usage{
harmonize_field_names(x = NULL)
}
\arguments{
\item{x}{a vector of field codes}
}
\value{
Vector of field names
}
\description{
Convert field codes to names
}
\details{
For MARC field descriptions, see \url{https://www.loc.gov/marc/bibliographic/}.
}
\examples{
x2 <- harmonize_field_names("041a")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
