% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/harmonize_ie.R
\name{harmonize_ie}
\alias{harmonize_ie}
\title{harmonize_ie}
\usage{
harmonize_ie(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A vector polished
}
\description{
Harmonize ie statement
}
\examples{
\dontrun{x2 <- harmonize_ie("i.e.")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

