% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_ie.R
\name{harmonize_ie}
\alias{harmonize_ie}
\title{Harmonize ie}
\usage{
harmonize_ie(x, separator = "i.e")
}
\arguments{
\item{x}{A vector}

\item{separator}{The separator string (i.e by default)}
}
\value{
A vector polished
}
\description{
Harmonize ie statements.
}
\examples{
\dontrun{x2 <- harmonize_ie("i.e.")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
