% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_print_statements.R
\name{harmonize_print_statements}
\alias{harmonize_print_statements}
\title{Harmonize Print Statements}
\usage{
harmonize_print_statements(x, lowercase = FALSE)
}
\arguments{
\item{x}{a vector}

\item{lowercase}{Set all in lowercase}
}
\value{
Harmonized vector
}
\description{
Harmonize print statements.
}
\examples{
# x2 <- harmonize_print_statements("prentyd in London")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
