% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonize_publisher.R
\name{harmonize_publisher}
\alias{harmonize_publisher}
\title{Harmonize Publisher}
\usage{
harmonize_publisher(x, publication_year, languages = c("english"))
}
\arguments{
\item{x}{A vector of publisher names}

\item{publication_year}{Data frame with "from" and "till" years}

\item{languages}{A vector of languages which are used in detecting relation keywords}
}
\value{
Data frame with orig, mod and comp
}
\description{
Harmonizes publishers with slight misspellings.
}
\examples{
# harmonize_publisher(x, publication_year, languages=c("finnish", "swedish", "latin"))
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
