% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.issue.R
\name{is.issue}
\alias{is.issue}
\title{Issue Identification}
\usage{
is.issue(df)
}
\arguments{
\item{df}{data.frame of documents x variables}
}
\value{
Logical vector indicating the issues
}
\description{
Identify documents that can be considered issues; based on other document info.
}
\details{
This function is used only to estimate pagecounts for documents with missing page count information.
Therefore no page count is considered in assessing the issue status.
}
\examples{
\dontrun{issue <- is.issue(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
