% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.multivol.R
\name{is.multivol}
\alias{is.multivol}
\title{Identify Multi-Volume Documents}
\usage{
is.multivol(df)
}
\arguments{
\item{df}{data.frame of documents x variables}
}
\value{
Logical vector indicating the multi-volume docs
}
\description{
Identify multi-volume documents based on the overall data.
}
\details{
Used to estimate pagecounts for documents with missing page count information. Therefore no page count is considered in assessing the issue status.
}
\examples{
\dontrun{issue <- is.multivol(df)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

