% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map.R
\name{map}
\alias{map}
\title{Map Terms}
\usage{
map(
  x,
  synonymes,
  remove.unknown = FALSE,
  mode = "exact.match",
  verbose = FALSE,
  from = "synonyme",
  to = "name",
  keep.names = FALSE
)
}
\arguments{
\item{x}{A character vector}

\item{synonymes}{synonyme table with the fields 'synonyme' and 'name'}

\item{remove.unknown}{Logical. Remove terms that do not have synonymes.}

\item{mode}{'exact.match' replaces the terms based on the synonyme list
if an exact match is found; 'match' replaces the parts that match
synonymes; 'recursive' replaces all (sub)strings recursively in
the same order as in the synonyme table}

\item{verbose}{verbose}

\item{from}{field that will be replaced}

\item{to}{field that contains the final names}

\item{keep.names}{Keep the original names in the returned vector
(may slow down analysis in large data sets)}
}
\value{
Vector of harmonized terms
}
\description{
Map between two sets of terms; used to harmonize data.
}
\details{
Map the input vector to harmonized versions based on the synonyme
         table.
}
\examples{
\dontrun{x2 <- map(x, synonymes)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
