% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_duplicates.R
\name{mark_duplicates}
\alias{mark_duplicates}
\title{Mark Duplicates}
\usage{
mark_duplicates(df.preprocessed, field_names = c("short_title",
  "publication_place", "publication_year"))
}
\arguments{
\item{df.preprocessed}{Data frame of combined catalogs}

\item{field_names}{List of field names of exactly matching values}
}
\value{
Data frame
}
\description{
Tag entries to be removed.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

