% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_languages.R
\name{mark_languages}
\alias{mark_languages}
\title{Mark Languages}
\usage{
mark_languages(x)
}
\arguments{
\item{x}{language field (a vector)}
}
\value{
data.frame with separate fields for different languages
}
\description{
Construct binary matrix of languages for each entry
}
\examples{
\dontrun{df <- mark_languages(c("fin;lat","eng"))}
}
\references{
See citation("bibliographica")
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\keyword{utilities}
