% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_pagecounts.R
\name{mean_pagecounts}
\alias{mean_pagecounts}
\title{Mean Page Counts}
\usage{
mean_pagecounts(df, pagecount.field = "pagecount", rounding = FALSE)
}
\arguments{
\item{df}{data frame, including the required fields.}

\item{pagecount.field}{Field to be used in estimation.}

\item{rounding}{Round to the closest integer}
}
\value{
Average page count.
}
\description{
Estimate mean and median page counts.
}
\examples{
\dontrun{m <- mean_pagecounts(df)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
