% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_gatherings.R
\name{order_gatherings}
\alias{order_gatherings}
\title{order_gatherings}
\usage{
order_gatherings(x)
}
\arguments{
\item{x}{A vector or factor of gathering data}
}
\value{
A factor with ordered gatherings levels
}
\description{
Sort gatherings levels
}
\examples{
order_gatherings(factor(c("2fo", "1to", "8vo")))
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
