% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_firstname.R
\name{pick_firstname}
\alias{pick_firstname}
\title{Pick First Name}
\usage{
pick_firstname(x, format = "last, first", keep.single = FALSE)
}
\arguments{
\item{x}{a vector of full names}

\item{format}{name format}

\item{keep.single}{If the name is without comma ('Shakespeare,
William' versus 'William'), interpret the name as first name.
Note that in this case also 'Shakespeare' will be interpreted as first name.}
}
\value{
a vector of first names
}
\description{
Pick first name from full name, assuming the format is known
}
\examples{
pick_firstname("Hobbes, Thomas")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
