% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_lastname.R
\name{pick_lastname}
\alias{pick_lastname}
\title{Pick Last Name}
\usage{
pick_lastname(x, format = "last, first", keep.single = TRUE)
}
\arguments{
\item{x}{a vector of full names}

\item{format}{name format}

\item{keep.single}{If the name is without comma ('Shakespeare,
William' versus 'William'), interpret the name as first name.
Note that in this case also 'Shakespeare' will be interpreted as first name.
In the current implementation keep.single is always TRUE.}
}
\value{
a vector of last names
}
\description{
Pick last name from full name, assuming the format is known
}
\examples{
pick_lastname("Hobbes, Thomas")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
