% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_multivolume.R
\name{pick_multivolume}
\alias{pick_multivolume}
\title{Pick multivolume}
\usage{
pick_multivolume(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Volume information
}
\description{
Pick volume information for multi-volume documents.
}
\examples{
\dontrun{pick_multivolume("v.1-3, 293")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
