% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_all.R
\name{polish_all}
\alias{polish_all}
\title{Polish all fields}
\usage{
polish_all(df.orig, fields = NULL, verbose = TRUE, file = NULL)
}
\arguments{
\item{df.orig}{Original data.frame}

\item{fields}{Fields to be preprocessed.}

\item{verbose}{verbose}

\item{file}{Temporary results saved into a file}
}
\value{
Preprocessed data.frame and field conversion list.
}
\description{
Polish all fields.
}
\examples{
\dontrun{a <- polish_all(df.orig)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

