% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_dimension.R
\name{polish_dimension}
\alias{polish_dimension}
\title{Polish Dimension}
\usage{
polish_dimension(x, synonyms)
}
\arguments{
\item{x}{A dimension note (a single string of one document)}

\item{synonyms}{synonyms}
}
\value{
Polished dimension information with the original string 
	   and gatherings and cm information collected from it
}
\description{
Polish dimension field of a single document
}
\examples{
# polish_dimension("4")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}
