% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_dimensions.R
\name{polish_dimensions}
\alias{polish_dimensions}
\title{polish_dimensions}
\usage{
polish_dimensions(x, fill = FALSE, dimtab = NULL)
}
\arguments{
\item{x}{A vector of dimension notes}

\item{fill}{Logical. Estimate and fill in the missing information: TRUE/FALSE}

\item{dimtab}{Dimension mapping table}
}
\value{
Dimension table
}
\description{
Polish dimension field for many documents at once
}
\examples{
polish_dimensions(c("2fo", "14cm"), fill = TRUE)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

