% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_field.R
\name{polish_field}
\alias{polish_field}
\title{Polish field}
\usage{
polish_field(df, field, verbose = TRUE)
}
\arguments{
\item{df}{data.frame that includes the given field}

\item{field}{Field to be preprocessed.}

\item{verbose}{verbose}
}
\value{
Output of the polished field
}
\description{
Polish a specified library catalogue field.
}
\details{
A wrapper bringing together the separate field processing methods into a single function.
}
\examples{
\dontrun{a <- polish_field(df, "title")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

