% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_name_of_author.R
\name{polish_name_of_author}
\alias{polish_name_of_author}
\title{polish_name_of_author}
\usage{
polish_name_of_author(x)
}
\arguments{
\item{x}{author name field (a vector)}
}
\value{
data.frame with fields for family and first name if available
}
\description{
Pick and polish name of author
}
\examples{
\dontrun{df <- polish_name_of_author(c("Washington, George","Louis XIV"))}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

