% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_pages.R
\name{polish_page}
\alias{polish_page}
\title{polish_page}
\usage{
polish_page(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Cleaned up version of page number field.
}
\description{
Clean up page numbers for a single document.
}
\examples{
\dontrun{polish_page("4p.")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{internal}

