% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_pages.R
\name{polish_pages}
\alias{polish_pages}
\title{polish_pages}
\usage{
polish_pages(x, verbose = FALSE)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}

\item{verbose}{Print progress info}
}
\value{
Raw and estimated pages per document part
}
\description{
clean up page numbers
}
\details{
Document parts are separated by semicolons

A summary of page counting rules that this function aims to (approximately) implement are provided in 
\url{https://www.libraries.psu.edu/psul/cataloging/training/bpcr/300.html}
}
\examples{
# polish_pages("4p.")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

