% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_physical_extent.R
\name{polish_physext_help2}
\alias{polish_physext_help2}
\title{Polish physical_extent help field 2}
\usage{
polish_physext_help2(x, page.harmonize)
}
\arguments{
\item{x}{Input char}
}
\value{
Internal
}
\description{
Internal
}
\examples{
# TBA
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{internal}
