% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_place.R
\name{polish_place}
\alias{polish_place}
\title{Polish Place}
\usage{
polish_place(
  x,
  synonymes = NULL,
  remove.unknown = FALSE,
  verbose = FALSE,
  harmonize = TRUE
)
}
\arguments{
\item{x}{A vector of place names}

\item{synonymes}{Synonyme table for place names}

\item{remove.unknown}{Logical. Remove places that are not validated (ie. listed in the synonyme table)?}

\item{verbose}{verbose}

\item{harmonize}{Harmonize the polished names. Logical.}
}
\value{
Polished vector
}
\description{
Polish place names.
}
\examples{
# x2 <- polish_place(c("London", "Paris"))
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
