% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{Polish Publisher}
\usage{
polish_publisher(x, verbose = TRUE, mc.cores = 1)
}
\arguments{
\item{x}{publisher field (a vector)}

\item{verbose}{verbose}

\item{mc.cores}{Number of cores for parallelization}
}
\value{
polished publisher field (a vector)
}
\description{
Generic function for preliminary polishing of publishing house names.
}
\examples{
\dontrun{v <- polish_publisher(c("Oxford University Press","tryckt hos Cambridge University Press"))}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

