% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{Generic Cleanup for the Publisher Field}
\usage{
polish_publisher(df, languages = "english")
}
\arguments{
\item{df}{Corresponding data frame assuming the place, year fields are already polished and available.}

\item{languages}{languages to consider in polishing}
}
\value{
A vector with polished entries.
}
\description{
Main handler for publisher fields.
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

