% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{Polish Publisher Main}
\usage{
polish_publisher(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
Data frame with orig, mod
}
\description{
Main handler for publisher fields.
}
\author{
Leo Lahti \email{hege.roivainen@gmail.com}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

