% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{polish_publisher}
\usage{
polish_publisher(x)
}
\arguments{
\item{x}{publisher field (a vector)}
}
\value{
polished publisher field (a vector)
}
\description{
Polish publishing house names
}
\examples{
\dontrun{v <- polish_publisher(c("Oxford University Press","tryckt hos Cambridge University Press"))}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

