% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_publisher.R
\name{polish_publisher}
\alias{polish_publisher}
\title{Polish Publisher Generic}
\usage{
polish_publisher(x)
}
\arguments{
\item{x}{Character vector of publisher names}
}
\value{
Data frame with orig, mod
}
\description{
Generic cleanup of the publisher field.
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
