% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_title.R
\name{polish_title}
\alias{polish_title}
\title{Polish Title}
\usage{
polish_title(x)
}
\arguments{
\item{x}{Vector of titles}
}
\value{
Vector of titles polished
}
\description{
Polish the title field.
}
\details{
Remove ending commas, periods, spaces and parentheses, 
	    starting prepositions etc.
}
\examples{
\dontrun{x2 <- polish_title(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
