% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_university.R
\name{polish_university}
\alias{polish_university}
\title{Polish University}
\usage{
polish_university(x, synonyms = NULL)
}
\arguments{
\item{x}{university field (a vector)}

\item{synonyms}{Synonyme table}
}
\value{
polished university field (a vector)
}
\description{
Polish university names
}
\examples{
\dontrun{v <- polish_university(c("HY","Suomen yliopisto"))}
}
\references{
See citation("bibliographica")
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\keyword{utilities}
