% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/volume.R
\name{polish_volumenumber}
\alias{polish_volumenumber}
\title{polish_volumenumber}
\usage{
polish_volumenumber(s)
}
\arguments{
\item{s}{Page number field. Vector or factor of strings.}
}
\value{
Volume number
}
\description{
Get volume number from page field if available
}
\details{
Refers to single-volume document where the volume has been specified
}
\examples{
polish_volumenumber("v.4")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

