% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_volumenumber.R
\name{polish_volumenumber}
\alias{polish_volumenumber}
\title{polish_volumenumber}
\usage{
polish_volumenumber(s, harmonize_volume = TRUE)
}
\arguments{
\item{s}{Page number field. Vector or factor of strings.}

\item{harmonize_volume}{Harmonize volume info. Logical.}
}
\value{
Volume number
}
\description{
Get volume number from page field if available
}
\details{
Refers to single-volume document where the volume has been specified
}
\examples{
polish_volumenumber("v.4")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{?}
\keyword{A}
\keyword{but}
\keyword{check}
\keyword{document,}
\keyword{single}
\keyword{utilities}
\keyword{volume}
\keyword{which}

