% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_years.R
\name{polish_year}
\alias{polish_year}
\title{polish_year}
\usage{
polish_year(x, start_synonyms = NULL, end_synonyms = NULL)
}
\arguments{
\item{x}{year string}

\item{start_synonyms}{Synonyme table for start year}

\item{end_synonyms}{Synonyme table for end year}
}
\value{
vector with the fields 'from' and 'till'
}
\description{
Pick and polish the year interval (start and end
}
\examples{
\dontrun{df <- polish_year(c("1746"))}
}
\author{
Leo Lahti and Niko Ilomaki \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

