% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polish_year_of_publication.R
\name{polish_year_of_publication}
\alias{polish_year_of_publication}
\title{polish_year_of_publication}
\usage{
polish_year_of_publication(x)
}
\arguments{
\item{x}{publication time field (a vector)}
}
\value{
data.frame for separate fields for books (year) and magazines (start, end)
}
\description{
Pick and polish year(s) of publication
}
\examples{
\dontrun{df <- polish_year_of_publication(c("1776","1803-1806"))}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

