% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polish_years.R
\name{polish_years}
\alias{polish_years}
\title{polish_years}
\usage{
polish_years(x)
}
\arguments{
\item{x}{year field (a vector)}
}
\value{
data.frame with the fields 'start' and 'end'
}
\description{
Pick and polish the year interval (start and end
}
\examples{
\dontrun{df <- polish_years("1746", "1745-1750")}
}
\author{
Niko Ilomaki \email{niko.ilomaki@helsinki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

