% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publication_frequency_text.R
\name{publication_frequency_text}
\alias{publication_frequency_text}
\title{Annual Publication Frequency to Text}
\usage{
publication_frequency_text(x, peryear)
}
\arguments{
\item{x}{Original publication frequency text}

\item{peryear}{Estimated annual publication frequency}
}
\description{
Convert annual publication frequencies to text format.
}
\examples{
\dontrun{df <- publication_frequency_text("Twice per year" 2)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
