% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bibliographic_metadata.R
\name{read_bibliographic_metadata}
\alias{read_bibliographic_metadata}
\title{Read Bibliographic Metadata}
\usage{
read_bibliographic_metadata(file, verbose = FALSE, sep = "|")
}
\arguments{
\item{file}{Parsed raw data file/s}

\item{verbose}{verbose}

\item{sep}{separator}
}
\value{
data.frame with raw data fields
}
\description{
Read metadata parsed from XML.
}
\examples{
\dontrun{df.raw <- read_bibliographic_metadata(file)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
