% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_dimension_info.R
\name{remove_dimension_info}
\alias{remove_dimension_info}
\title{Remove dimension information from a single document}
\usage{
remove_dimension_info(x)
}
\arguments{
\item{x}{A character vector that may contain dimension information}
}
\value{
The character vector with dimension information removed
}
\description{
Remove dimension information from a single document
}
\examples{
remove_dimension_info("4to 40cm")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{internal}

