% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_duplicates.R
\name{remove_duplicates}
\alias{remove_duplicates}
\title{Removes duplicates}
\usage{
remove_duplicates(
  df.preprocessed,
  field_names = c("short_title", "publication_place", "publication_year")
)
}
\arguments{
\item{df.preprocessed}{Data frame of combined catalogs}

\item{field_names}{List of field names of exactly matching values}
}
\value{
Data frame
}
\description{
Tags files as to be removed
}
\references{
See citation("bibliographica")
}
\author{
Hege Roivainen \email{hege.roivainen@gmail.com}
}
\keyword{utilities}
