% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_letters.R
\name{remove_letters}
\alias{remove_letters}
\title{Remove Letters.}
\usage{
remove_letters(
  x,
  ignore.start = c("S", "s", "N", "n", "E", "e", "W", "w"),
  ignore.end = c()
)
}
\arguments{
\item{x}{A vector}

\item{ignore.start}{Vector of letter not to exclude from the start of x}

\item{ignore.end}{Vector of letter not to exclude from the end of x}
}
\value{
Polished vector
}
\description{
Remove specific single letters from the start and/or end of the input string(s).
}
\examples{
\dontrun{remove_letters(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
