% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/remove_persons.R
\name{remove_persons}
\alias{remove_persons}
\title{remove_persons}
\usage{
remove_persons(x, persons = NULL)
}
\arguments{
\item{x}{A vector}

\item{persons}{names to be removed}
}
\value{
Polished vector
}
\description{
Remove persons
}
\examples{
\dontrun{x2 <- remove_persons(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

