% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_persons.R
\name{remove_persons}
\alias{remove_persons}
\title{Remove Persons}
\usage{
remove_persons(x, who = NULL)
}
\arguments{
\item{x}{A vector}

\item{who}{names to be removed}
}
\value{
Polished vector
}
\description{
Remove persons.
}
\examples{
\dontrun{x2 <- remove_persons(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
