% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_print_statements.R
\name{remove_print_statements}
\alias{remove_print_statements}
\title{Remove Print Statements}
\usage{
remove_print_statements(x)
}
\arguments{
\item{x}{a vector}
}
\value{
Polished vector
}
\description{
Remove print statements.
}
\examples{
x2 <- remove_print_statements("Printed in London")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
