% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_special_chars}
\alias{remove_special_chars}
\title{remove_special_chars}
\usage{
remove_special_chars(x, chars = c(",", ";", ":", "\\\\(", "\\\\)", "\\\\?",
  "--", "\\\\&"), niter = 5)
}
\arguments{
\item{x}{Character vector}

\item{chars}{Characters to be removed}

\item{niter}{Number of iterations}
}
\value{
Polished vector
}
\description{
Remove special characters
}
\examples{
\dontrun{remove_special_chars()}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

