% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_special_chars.R
\name{remove_special_chars}
\alias{remove_special_chars}
\title{Remove Special Chars}
\usage{
remove_special_chars(
  x,
  chars = c(",", ";", ":", "\\\\(", "\\\\)", "\\\\?", "--", "\\\\&"),
  niter = 5
)
}
\arguments{
\item{x}{Character vector}

\item{chars}{Characters to be removed}

\item{niter}{Number of iterations}
}
\value{
Polished vector
}
\description{
Remove special characters.
}
\examples{
x2 <- remove_special_chars("test;")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
