% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/stopwords.R
\name{remove_stopwords}
\alias{remove_stopwords}
\title{remove_stopwords}
\usage{
remove_stopwords(x, terms = NULL, remove.letters = TRUE)
}
\arguments{
\item{x}{A vector}

\item{terms}{Stopwords}

\item{remove.letters}{Logical. Also remove single letters (TRUE by default)}
}
\value{
A vector polished
}
\description{
Remove stopwords from input vector
}
\examples{
x2 <- remove_stopwords(c("a", "well", "james"), "well", remove.letters = TRUE)
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

