% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{remove_terms}
\alias{remove_terms}
\title{remove_terms}
\usage{
remove_terms(x, terms, where = "all")
}
\arguments{
\item{x}{A vector}

\item{terms}{Terms to be removed}

\item{where}{Locations to be removed ("all" / "begin" / "middle" / "end")}
}
\value{
Vector with terms removed
}
\description{
Remove the given terms from the strings
}
\details{
After removing the numerics, beginning, double and ending
         spaces are also removed from the strings.
}
\examples{
\dontrun{x2 <- remove_terms(x, terms, where = "all")}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

