% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_terms.R
\name{remove_terms}
\alias{remove_terms}
\title{Remove Terms}
\usage{
remove_terms(
  x,
  terms,
  where = "all",
  include.lowercase = FALSE,
  polish = TRUE,
  recursive = FALSE
)
}
\arguments{
\item{x}{A vector}

\item{terms}{Terms to be removed}

\item{where}{Locations to be removed ("all" / "begin" / "middle" / "end")}

\item{include.lowercase}{Include also lowercase versions of the terms}

\item{polish}{polish the entries after removing the terms (remove trailing spaces and periods)}

\item{recursive}{Apply the changes recursively along the list ?}
}
\value{
Vector with terms removed
}
\description{
Remove the given terms from the strings.
}
\details{
After removing the numerics, beginning, double and ending 
         spaces are also removed from the strings.
}
\examples{
x2 <- remove_terms("test this", c("this", "works"), where = "all")
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
