% Generated by roxygen2 (4.0.2): do not edit by hand
\name{remove_trailing_periods}
\alias{remove_trailing_periods}
\title{remove_trailing_periods}
\usage{
remove_trailing_periods(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A polished vector
}
\description{
Remove trailing periods
}
\examples{
\dontrun{x2 <- remove_trailing_periods(x)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

