% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_trailing_periods.R
\name{remove_trailing_periods}
\alias{remove_trailing_periods}
\title{Remove Trailing Periods}
\usage{
remove_trailing_periods(x)
}
\arguments{
\item{x}{A vector}
}
\value{
A polished vector
}
\description{
Remove trailing periods.
}
\examples{
\dontrun{x2 <- remove_trailing_periods(x)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
