% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_volume_info.R
\name{remove_volume_info}
\alias{remove_volume_info}
\title{Remove Volume Info}
\usage{
remove_volume_info(x)
}
\arguments{
\item{x}{Page number field. Vector or factor of strings.}
}
\value{
Page numbers without volume information
}
\description{
Remove volume info from the string start.
}
\examples{
\dontrun{remove_volume_info("v.4, 293")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
