% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheet_area.R
\name{sheet_area}
\alias{sheet_area}
\title{sheet_area}
\usage{
sheet_area(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{Sheet size}

\item{verbose}{Verbose}
}
\value{
Sheet area (cm2)
}
\description{
Sheet area in cm2
}
\details{
Sheet size is calculated according to the table given as output from call sheet_area()
}
\examples{
sheet_area("2to")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

