% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ambiguous_country.R
\name{test_ambiguous_country}
\alias{test_ambiguous_country}
\title{Identify Ambiguous Country Mappings}
\usage{
test_ambiguous_country(
  f = system.file("extdata/reg2country.csv", package = "bibliographica")
)
}
\arguments{
\item{f}{Input file}
}
\value{
List of ambiguous country mappings
}
\description{
Identify ambiguous country mappings in the region-counry mapping table.
}
\examples{
aa <- test_ambiguous_country()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
