% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_ambiguous_place.R
\name{test_ambiguous_place}
\alias{test_ambiguous_place}
\title{Identify Ambiguous Place Mappings}
\usage{
test_ambiguous_place(
  f = system.file("extdata/PublicationPlaceSynonymes.csv", package = "bibliographica")
)
}
\arguments{
\item{f}{Input file}
}
\value{
List of ambiguous place mappings
}
\description{
Identify ambiguous place mappings in the region-counry mapping table.
}
\examples{
aa <- test_ambiguous_place()
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
