% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline.R
\name{timeline}
\alias{timeline}
\title{Retrieve Timeline}
\usage{
timeline(
  x,
  field = "titlecount",
  group = NULL,
  nmin = 0,
  mode = "absolute",
  time.window = 10,
  time.field = "publication_year"
)
}
\arguments{
\item{x}{data frame}

\item{field}{Numeric field to summarize in the timeline. The number of entries (title count) per decade is used by default. If this argument is used, the sum of entries per decade for this field is given.}

\item{group}{Optional. Name for a data field that indicates groups to compare.}

\item{nmin}{Include only entries with at least nmin absolute frequency}

\item{mode}{"absolute" or "relative"}

\item{time.window}{Time window for the timeline in years. Default: 10 (publication decade).}

\item{time.field}{Specify the field to be used for time. By default: "publication_year", or if time.window is 10, then "publication_decade"}
}
\value{
data.frame
}
\description{
Timeline data for selected variable (possibly across various groups).
}
\examples{
\dontrun{timeline(df, "gatherings")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
