% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/top_plot.R
\name{top_plot}
\alias{top_plot}
\title{top_plot}
\usage{
top_plot(df, field, ntop = 20)
}
\arguments{
\item{df}{Data frame}

\item{field}{Field to show}

\item{ntop}{Number of top entries to show}
}
\value{
ggplot object
}
\description{
Plot the top entries for a given field in a data frame
}
\examples{
\dontrun{p <- top_plot(df, field, 50)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

