% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_plot.R
\name{top_plot}
\alias{top_plot}
\title{Plot top entries}
\usage{
top_plot(df, field = NULL, ntop = NULL, highlight = NULL)
}
\arguments{
\item{df}{Data frame, vector or factor}

\item{field}{Field to show}

\item{ntop}{Number of top entries to show}

\item{highlight}{Entries from the 'field' to be highlighted}
}
\value{
ggplot object
}
\description{
Plot the top entries for a given field in a data frame
}
\examples{
\dontrun{p <- top_plot(df, field, 50)}
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\references{
See citation("bibliographica")
}
\keyword{utilities}

