% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming.R
\name{trimming}
\alias{trimming}
\title{Trimming}
\usage{
trimming(x, n = 1)
}
\arguments{
\item{x}{vector}

\item{n}{iterations}
}
\value{
Polished vector
}
\description{
Trim strings.
}
\examples{
\dontrun{x2 <- trimming(x, n = 1)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
