% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_preprocessed_data_file.R
\name{update_preprocessed_data_file}
\alias{update_preprocessed_data_file}
\title{Update Preprocessed Data File}
\usage{
update_preprocessed_data_file(file, df.preprocessed, mode = "update")
}
\arguments{
\item{file}{File name}

\item{df.preprocessed}{Preprocessed data.frame}

\item{mode}{"update" or "overwrite"; the former updates the given fields;
the latter overwrites the file completely}
}
\value{
File name
}
\description{
Update Preprocessed Data File.
}
\examples{
\dontrun{f <- update_preprocessed_data_file(file)}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
