% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_names.R
\name{validate_names}
\alias{validate_names}
\title{Validate Names}
\usage{
validate_names(df.preprocessed)
}
\arguments{
\item{df.preprocessed}{Preprocessed data.frame.}
}
\value{
Modified data.frame
}
\description{
Validation for names.
}
\examples{
# \dontrun{validate_names(df.preprocessed)}
}
\references{
See citation("bibliographica")
}
\author{
Ville Vaara and Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
