% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_xtable.R
\name{write_xtable}
\alias{write_xtable}
\title{Write Summary Table}
\usage{
write_xtable(
  x,
  filename = NULL,
  count = FALSE,
  sort.by = "Count",
  na.rm = TRUE,
  add.percentages = FALSE
)
}
\arguments{
\item{x}{a vector or matrix}

\item{filename}{output file}

\item{count}{Add total count of cases in the beginning}

\item{sort.by}{Column used for sorting. The Count is the default.}

\item{na.rm}{Remove NA entries}

\item{add.percentages}{Add percentage information to the table. This indicates the total fraction of the count, calculated from all input entries if na.rm is FALSE, and from non-NA entries if na.rm is TRUE.}
}
\value{
Table indicating the count for each unique entry in the input  vector or matrix. The function writes the statistics in the file.
}
\description{
Write xtable in a file
}
\examples{
\dontrun{tab <- write_xtable(x, "tmp.tab")}
}
\references{
See citation("bibliographica")
}
\author{
Leo Lahti \email{leo.lahti@iki.fi}
}
\keyword{utilities}
